Add the coordinates of SP on front and use on heatmap
