--- src/components/HeatMap/coordinates.ts
@@ -1,1996 +1,0 @@
-export const coordinatesGama = [
-    { longitude: -48.005184173584, latitude: -15.9778680801392 },
-    { longitude: -48.0125694274901, latitude: -15.9958820343018 },
-    { longitude: -48.0249176025391, latitude: -16.0073585510253 },
-    { longitude: -48.0387496948242, latitude: -16.0193901062012 },
-    { longitude: -48.0444183349608, latitude: -16.0197982788086 },
-    { longitude: -48.0544242858886, latitude: -16.0375995635986 },
-    { longitude: -48.0470314025879, latitude: -16.0371894836425 },
-    { longitude: -48.034366607666, latitude: -16.0516242980957 },
-    { longitude: -48.0866317749023, latitude: -16.0517616271972 },
-    { longitude: -48.2790756225586, latitude: -16.0515556335449 },
-    { longitude: -48.2787055969238, latitude: -16.0488243103027 },
-    { longitude: -48.2715797424316, latitude: -16.028709411621 },
-    { longitude: -48.2728042602538, latitude: -16.0084781646729 },
-    { longitude: -48.2648239135742, latitude: -15.9948043823242 },
-    { longitude: -48.260196685791, latitude: -15.9777631759643 },
-    { longitude: -48.2421417236328, latitude: -15.9714183807372 },
-    { longitude: -48.2275657653809, latitude: -15.9718303680419 },
-    { longitude: -48.2023544311523, latitude: -15.9633979797363 },
-    { longitude: -48.1874046325684, latitude: -15.9639701843262 },
-    { longitude: -48.1846389770508, latitude: -15.9691228866577 },
-    { longitude: -48.1754837036133, latitude: -15.9569816589355 },
-    { longitude: -48.1706085205078, latitude: -15.9399356842041 },
-    { longitude: -48.1585388183593, latitude: -15.9371213912963 },
-    { longitude: -48.1568031311035, latitude: -15.9386987686157 },
-    { longitude: -48.1437377929688, latitude: -15.9404420852661 },
-    { longitude: -48.1431732177734, latitude: -15.9507551193237 },
-    { longitude: -48.1344833374023, latitude: -15.9549207687378 },
-    { longitude: -48.1232910156249, latitude: -15.9763889312744 },
-    { longitude: -48.1194686889648, latitude: -15.975438117981 },
-    { longitude: -48.1252326965332, latitude: -15.9660711288452 },
-    { longitude: -48.1240921020508, latitude: -15.9568977355956 },
-    { longitude: -48.113166809082, latitude: -15.9516258239746 },
-    { longitude: -48.0912818908691, latitude: -15.9264354705809 },
-    { longitude: -48.0828094482422, latitude: -15.9275970458984 },
-    { longitude: -48.082893371582, latitude: -15.9360656738281 },
-    { longitude: -48.0500068664551, latitude: -15.953408241272 },
-    { longitude: -48.0427169799805, latitude: -15.9513177871704 },
-    { longitude: -48.0375900268555, latitude: -15.9452848434448 },
-    { longitude: -48.0367927551268, latitude: -15.9462537765503 },
-    { longitude: -48.0210342407226, latitude: -15.9652757644653 },
-    { longitude: -48.005184173584, latitude: -15.9778680801392 },
-];
-
-export const coordinatesGuara = [
-    { longitude: -48.020000000000003, latitude: -15.826000000000001 },
-    { longitude: -48.024000000000001, latitude: -15.823 },
-    { longitude: -48.024000000000001, latitude: -15.819000000000001 },
-    { longitude: -48.009999999999998, latitude: -15.812000000000001 },
-    { longitude: -48.011000000000003, latitude: -15.81 },
-    { longitude: -48.009999999999998, latitude: -15.790000000000001 },
-    { longitude: -48.006999999999998, latitude: -15.784000000000001 },
-    { longitude: -48.003999999999998, latitude: -15.782 },
-    { longitude: -47.988, latitude: -15.782999999999999 },
-    { longitude: -47.984999999999999, latitude: -15.787000000000001 },
-    { longitude: -47.988, latitude: -15.803000000000001 },
-    { longitude: -47.980000000000004, latitude: -15.801 },
-    { longitude: -47.971000000000004, latitude: -15.803000000000001 },
-    { longitude: -47.968000000000004, latitude: -15.805 },
-    { longitude: -47.966000000000001, latitude: -15.815 },
-    { longitude: -47.957999999999998, latitude: -15.812000000000001 },
-    { longitude: -47.956000000000003, latitude: -15.814 },
-    { longitude: -47.951999999999998, latitude: -15.814 },
-    { longitude: -47.945999999999998, latitude: -15.819000000000001 },
-    { longitude: -47.951999999999998, latitude: -15.854000000000001 },
-    { longitude: -47.954000000000001, latitude: -15.858000000000001 },
-    { longitude: -47.957999999999998, latitude: -15.858000000000001 },
-    { longitude: -47.954999999999998, latitude: -15.861000000000001 },
-    { longitude: -47.954999999999998, latitude: -15.865 },
-    { longitude: -47.960000000000001, latitude: -15.868 },
-    { longitude: -47.963999999999999, latitude: -15.868 },
-    { longitude: -47.968000000000004, latitude: -15.865 },
-    { longitude: -47.972000000000001, latitude: -15.866 },
-    { longitude: -47.975000000000001, latitude: -15.863 },
-    { longitude: -47.978999999999999, latitude: -15.862 },
-    { longitude: -47.981999999999999, latitude: -15.859 },
-    { longitude: -47.983000000000004, latitude: -15.855 },
-    { longitude: -47.987000000000002, latitude: -15.854000000000001 },
-    { longitude: -47.994, latitude: -15.847 },
-    { longitude: -47.997, latitude: -15.835000000000001 },
-    { longitude: -47.999000000000002, latitude: -15.834 },
-    { longitude: -48, latitude: -15.826000000000001 },
-    { longitude: -48.003, latitude: -15.823 },
-    { longitude: -48.009, latitude: -15.826000000000001 },
-    { longitude: -48.020000000000003, latitude: -15.826000000000001 },
-];
-
-export const coordinatesTaguatinga = [
-    { longitude: -48.116, latitude: -15.771000000000001 },
-    { longitude: -48.116, latitude: -15.769 },
-    { longitude: -48.113, latitude: -15.766 },
-    { longitude: -48.100000000000001, latitude: -15.761000000000001 },
-    { longitude: -48.094999999999999, latitude: -15.761000000000001 },
-    { longitude: -48.090000000000003, latitude: -15.757 },
-    { longitude: -48.084000000000003, latitude: -15.756 },
-    { longitude: -48.079000000000001, latitude: -15.753 },
-    { longitude: -48.07, latitude: -15.755000000000001 },
-    { longitude: -48.076999999999998, latitude: -15.74 },
-    { longitude: -48.076000000000001, latitude: -15.734 },
-    { longitude: -48.073999999999998, latitude: -15.733000000000001 },
-    { longitude: -48.061, latitude: -15.734 },
-    { longitude: -48.056000000000004, latitude: -15.736000000000001 },
-    { longitude: -48.048000000000002, latitude: -15.743 },
-    { longitude: -48.045000000000002, latitude: -15.749000000000001 },
-    { longitude: -48.017000000000003, latitude: -15.747 },
-    { longitude: -48.005000000000003, latitude: -15.75 },
-    { longitude: -48.002000000000002, latitude: -15.753 },
-    { longitude: -48.003, latitude: -15.769 },
-    { longitude: -48.002000000000002, latitude: -15.779 },
-    { longitude: -47.999000000000002, latitude: -15.784000000000001 },
-    { longitude: -47.999000000000002, latitude: -15.788 },
-    { longitude: -48.000999999999998, latitude: -15.790000000000001 },
-    { longitude: -48.020000000000003, latitude: -15.794 },
-    { longitude: -48.045999999999999, latitude: -15.795 },
-    { longitude: -48.048999999999999, latitude: -15.796000000000001 },
-    { longitude: -48.052, latitude: -15.804 },
-    { longitude: -48.050000000000004, latitude: -15.806000000000001 },
-    { longitude: -48.051000000000002, latitude: -15.812000000000001 },
-    { longitude: -48.048999999999999, latitude: -15.822000000000001 },
-    { longitude: -48.045999999999999, latitude: -15.828000000000001 },
-    { longitude: -48.044000000000004, latitude: -15.829000000000001 },
-    { longitude: -48.045000000000002, latitude: -15.834 },
-    { longitude: -48.033000000000001, latitude: -15.859 },
-    { longitude: -48.023000000000003, latitude: -15.875 },
-    { longitude: -48.024000000000001, latitude: -15.879 },
-    { longitude: -48.027999999999999, latitude: -15.880000000000001 },
-    { longitude: -48.048000000000002, latitude: -15.878 },
-    { longitude: -48.050000000000004, latitude: -15.875999999999999 },
-    { longitude: -48.050000000000004, latitude: -15.873000000000001 },
-    { longitude: -48.042999999999999, latitude: -15.867000000000001 },
-    { longitude: -48.047000000000004, latitude: -15.867000000000001 },
-    { longitude: -48.048999999999999, latitude: -15.865 },
-    { longitude: -48.054000000000002, latitude: -15.864000000000001 },
-    { longitude: -48.055, latitude: -15.861000000000001 },
-    { longitude: -48.058999999999997, latitude: -15.863 },
-    { longitude: -48.064999999999998, latitude: -15.862 },
-    { longitude: -48.073999999999998, latitude: -15.852 },
-    { longitude: -48.078000000000003, latitude: -15.85 },
-    { longitude: -48.079999999999998, latitude: -15.85 },
-    { longitude: -48.082999999999998, latitude: -15.854000000000001 },
-    { longitude: -48.087000000000003, latitude: -15.856 },
-    { longitude: -48.091999999999999, latitude: -15.855 },
-    { longitude: -48.097000000000001, latitude: -15.839 },
-    { longitude: -48.097000000000001, latitude: -15.836 },
-    { longitude: -48.093000000000004, latitude: -15.833 },
-    { longitude: -48.097000000000001, latitude: -15.824 },
-    { longitude: -48.097000000000001, latitude: -15.819000000000001 },
-    { longitude: -48.100999999999999, latitude: -15.812000000000001 },
-    { longitude: -48.116, latitude: -15.797000000000001 },
-    { longitude: -48.116, latitude: -15.792 },
-    { longitude: -48.113, latitude: -15.790000000000001 },
-    { longitude: -48.116, latitude: -15.771000000000001 },
-];
-
-export const coordinatesAguasClaras = [
-    { longitude: -48.048999999999999, latitude: -15.828000000000001 },
-    { longitude: -48.036000000000001, latitude: -15.826000000000001 },
-    { longitude: -48.021999999999998, latitude: -15.817 },
-    { longitude: -48.009, latitude: -15.817 },
-    { longitude: -48.005000000000003, latitude: -15.814 },
-    { longitude: -47.999000000000002, latitude: -15.814 },
-    { longitude: -47.997, latitude: -15.815 },
-    { longitude: -47.996000000000002, latitude: -15.819000000000001 },
-    { longitude: -47.994, latitude: -15.819000000000001 },
-    { longitude: -47.992000000000004, latitude: -15.822000000000001 },
-    { longitude: -47.989000000000004, latitude: -15.836 },
-    { longitude: -47.986000000000004, latitude: -15.839 },
-    { longitude: -47.986000000000004, latitude: -15.844000000000001 },
-    { longitude: -47.978999999999999, latitude: -15.847 },
-    { longitude: -47.974000000000004, latitude: -15.855 },
-    { longitude: -47.960999999999999, latitude: -15.859 },
-    { longitude: -47.957999999999998, latitude: -15.865 },
-    { longitude: -47.960000000000001, latitude: -15.868 },
-    { longitude: -47.983000000000004, latitude: -15.879 },
-    { longitude: -48.009999999999998, latitude: -15.882 },
-    { longitude: -48.029000000000003, latitude: -15.880000000000001 },
-    { longitude: -48.033000000000001, latitude: -15.875 },
-    { longitude: -48.053000000000004, latitude: -15.836 },
-    { longitude: -48.053000000000004, latitude: -15.83 },
-    { longitude: -48.048999999999999, latitude: -15.828000000000001 },
-];
-
-export const coordinatesCeilandia = [
-    { longitude: -48.289000000000001, latitude: -15.846 },
-    { longitude: -48.289000000000001, latitude: -15.835000000000001 },
-    { longitude: -48.286000000000001, latitude: -15.831 },
-    { longitude: -48.283999999999999, latitude: -15.821 },
-    { longitude: -48.277000000000001, latitude: -15.813000000000001 },
-    { longitude: -48.274999999999999, latitude: -15.813000000000001 },
-    { longitude: -48.274999999999999, latitude: -15.811 },
-    { longitude: -48.271999999999998, latitude: -15.809000000000001 },
-    { longitude: -48.268999999999998, latitude: -15.803000000000001 },
-    { longitude: -48.265000000000001, latitude: -15.803000000000001 },
-    { longitude: -48.258000000000003, latitude: -15.798 },
-    { longitude: -48.253, latitude: -15.798 },
-    { longitude: -48.250999999999998, latitude: -15.795 },
-    { longitude: -48.245000000000005, latitude: -15.793000000000001 },
-    { longitude: -48.245000000000005, latitude: -15.791 },
-    { longitude: -48.242000000000004, latitude: -15.789 },
-    { longitude: -48.242000000000004, latitude: -15.784000000000001 },
-    { longitude: -48.239000000000004, latitude: -15.781000000000001 },
-    { longitude: -48.237000000000002, latitude: -15.781000000000001 },
-    { longitude: -48.237000000000002, latitude: -15.777000000000001 },
-    { longitude: -48.233000000000004, latitude: -15.77 },
-    { longitude: -48.228000000000002, latitude: -15.768000000000001 },
-    { longitude: -48.228999999999999, latitude: -15.763 },
-    { longitude: -48.225000000000001, latitude: -15.76 },
-    { longitude: -48.221000000000004, latitude: -15.761000000000001 },
-    { longitude: -48.218000000000004, latitude: -15.749000000000001 },
-    { longitude: -48.216000000000001, latitude: -15.747 },
-    { longitude: -48.207000000000001, latitude: -15.748000000000001 },
-    { longitude: -48.189, latitude: -15.744 },
-    { longitude: -48.170000000000002, latitude: -15.747 },
-    { longitude: -48.158000000000001, latitude: -15.756 },
-    { longitude: -48.154000000000003, latitude: -15.764000000000001 },
-    { longitude: -48.133000000000003, latitude: -15.764000000000001 },
-    { longitude: -48.131, latitude: -15.766 },
-    { longitude: -48.122999999999998, latitude: -15.765000000000001 },
-    { longitude: -48.119, latitude: -15.766999999999999 },
-    { longitude: -48.111000000000004, latitude: -15.766 },
-    { longitude: -48.108000000000004, latitude: -15.769 },
-    { longitude: -48.103999999999999, latitude: -15.790000000000001 },
-    { longitude: -48.106999999999999, latitude: -15.795 },
-    { longitude: -48.103999999999999, latitude: -15.796000000000001 },
-    { longitude: -48.100999999999999, latitude: -15.802 },
-    { longitude: -48.097999999999999, latitude: -15.803000000000001 },
-    { longitude: -48.091999999999999, latitude: -15.811 },
-    { longitude: -48.088999999999999, latitude: -15.818 },
-    { longitude: -48.088999999999999, latitude: -15.823 },
-    { longitude: -48.084000000000003, latitude: -15.834 },
-    { longitude: -48.085000000000001, latitude: -15.839 },
-    { longitude: -48.088000000000001, latitude: -15.84 },
-    { longitude: -48.085000000000001, latitude: -15.855 },
-    { longitude: -48.105000000000004, latitude: -15.862 },
-    { longitude: -48.108000000000004, latitude: -15.870000000000001 },
-    { longitude: -48.112000000000002, latitude: -15.872 },
-    { longitude: -48.125, latitude: -15.872 },
-    { longitude: -48.131999999999998, latitude: -15.868 },
-    { longitude: -48.137999999999998, latitude: -15.868 },
-    { longitude: -48.152000000000001, latitude: -15.859 },
-    { longitude: -48.160000000000004, latitude: -15.859 },
-    { longitude: -48.161999999999999, latitude: -15.857000000000001 },
-    { longitude: -48.172000000000004, latitude: -15.859 },
-    { longitude: -48.176000000000002, latitude: -15.866 },
-    { longitude: -48.18, latitude: -15.866 },
-    { longitude: -48.183999999999997, latitude: -15.872 },
-    { longitude: -48.191000000000003, latitude: -15.877000000000001 },
-    { longitude: -48.200000000000003, latitude: -15.895 },
-    { longitude: -48.204000000000001, latitude: -15.896000000000001 },
-    { longitude: -48.209000000000003, latitude: -15.901 },
-    { longitude: -48.216000000000001, latitude: -15.902000000000001 },
-    { longitude: -48.218000000000004, latitude: -15.905000000000001 },
-    { longitude: -48.234000000000002, latitude: -15.906000000000001 },
-    { longitude: -48.236000000000004, latitude: -15.907999999999999 },
-    { longitude: -48.242000000000004, latitude: -15.909000000000001 },
-    { longitude: -48.245000000000005, latitude: -15.913 },
-    { longitude: -48.249000000000002, latitude: -15.913 },
-    { longitude: -48.250999999999998, latitude: -15.92 },
-    { longitude: -48.255000000000003, latitude: -15.922000000000001 },
-    { longitude: -48.253999999999998, latitude: -15.923 },
-    { longitude: -48.255000000000003, latitude: -15.931000000000001 },
-    { longitude: -48.256999999999998, latitude: -15.933 },
-    { longitude: -48.261000000000003, latitude: -15.933 },
-    { longitude: -48.265999999999998, latitude: -15.937000000000001 },
-    { longitude: -48.274000000000001, latitude: -15.935 },
-    { longitude: -48.277000000000001, latitude: -15.936 },
-    { longitude: -48.280000000000001, latitude: -15.934000000000001 },
-    { longitude: -48.280999999999999, latitude: -15.923 },
-    { longitude: -48.279000000000003, latitude: -15.919 },
-    { longitude: -48.274999999999999, latitude: -15.917 },
-    { longitude: -48.274999999999999, latitude: -15.914 },
-    { longitude: -48.280999999999999, latitude: -15.91 },
-    { longitude: -48.280999999999999, latitude: -15.891999999999999 },
-    { longitude: -48.277000000000001, latitude: -15.884 },
-    { longitude: -48.279000000000003, latitude: -15.883000000000001 },
-    { longitude: -48.280999999999999, latitude: -15.874000000000001 },
-    { longitude: -48.283999999999999, latitude: -15.871 },
-    { longitude: -48.285000000000004, latitude: -15.867000000000001 },
-    { longitude: -48.283999999999999, latitude: -15.863 },
-    { longitude: -48.288000000000004, latitude: -15.857000000000001 },
-    { longitude: -48.289000000000001, latitude: -15.846 },
-];
-
-export const coordinatesBrasilia = [
-    { longitude: -48.073999999999998, latitude: -15.74 },
-    { longitude: -48.084000000000003, latitude: -15.735000000000001 },
-    { longitude: -48.090000000000003, latitude: -15.727 },
-    { longitude: -48.094000000000001, latitude: -15.681000000000001 },
-    { longitude: -48.079999999999998, latitude: -15.630000000000001 },
-    { longitude: -48.078000000000003, latitude: -15.625999999999999 },
-    { longitude: -48.072000000000003, latitude: -15.621 },
-    { longitude: -48.048000000000002, latitude: -15.610000000000001 },
-    { longitude: -48.030000000000001, latitude: -15.588000000000001 },
-    { longitude: -48.020000000000003, latitude: -15.579000000000001 },
-    { longitude: -48.011000000000003, latitude: -15.574 },
-    { longitude: -48.003999999999998, latitude: -15.574 },
-    { longitude: -47.881, latitude: -15.665000000000001 },
-    { longitude: -47.856000000000002, latitude: -15.686 },
-    { longitude: -47.856000000000002, latitude: -15.69 },
-    { longitude: -47.858000000000004, latitude: -15.692 },
-    { longitude: -47.881999999999998, latitude: -15.704000000000001 },
-    { longitude: -47.895000000000003, latitude: -15.714 },
-    { longitude: -47.902000000000001, latitude: -15.724 },
-    { longitude: -47.895000000000003, latitude: -15.724 },
-    { longitude: -47.893000000000001, latitude: -15.726000000000001 },
-    { longitude: -47.889000000000003, latitude: -15.725 },
-    { longitude: -47.878999999999998, latitude: -15.726000000000001 },
-    { longitude: -47.870000000000005, latitude: -15.738 },
-    { longitude: -47.863, latitude: -15.74 },
-    { longitude: -47.859999999999999, latitude: -15.744 },
-    { longitude: -47.856999999999999, latitude: -15.744 },
-    { longitude: -47.843000000000004, latitude: -15.752000000000001 },
-    { longitude: -47.841000000000001, latitude: -15.757 },
-    { longitude: -47.832000000000001, latitude: -15.764000000000001 },
-    { longitude: -47.832000000000001, latitude: -15.76 },
-    { longitude: -47.834000000000003, latitude: -15.757 },
-    { longitude: -47.832999999999998, latitude: -15.752000000000001 },
-    { longitude: -47.835000000000001, latitude: -15.75 },
-    { longitude: -47.841999999999999, latitude: -15.748000000000001 },
-    { longitude: -47.847000000000001, latitude: -15.742000000000001 },
-    { longitude: -47.855000000000004, latitude: -15.737 },
-    { longitude: -47.863, latitude: -15.729000000000001 },
-    { longitude: -47.864000000000004, latitude: -15.725 },
-    { longitude: -47.861000000000004, latitude: -15.721 },
-    { longitude: -47.856999999999999, latitude: -15.720000000000001 },
-    { longitude: -47.844000000000001, latitude: -15.727 },
-    { longitude: -47.841999999999999, latitude: -15.731 },
-    { longitude: -47.837000000000003, latitude: -15.73 },
-    { longitude: -47.832000000000001, latitude: -15.735000000000001 },
-    { longitude: -47.832000000000001, latitude: -15.737 },
-    { longitude: -47.823, latitude: -15.741 },
-    { longitude: -47.817, latitude: -15.748000000000001 },
-    { longitude: -47.813000000000002, latitude: -15.749000000000001 },
-    { longitude: -47.811999999999998, latitude: -15.753 },
-    { longitude: -47.814999999999998, latitude: -15.758000000000001 },
-    { longitude: -47.810000000000002, latitude: -15.762 },
-    { longitude: -47.810000000000002, latitude: -15.766 },
-    { longitude: -47.814, latitude: -15.768000000000001 },
-    { longitude: -47.814, latitude: -15.77 },
-    { longitude: -47.813000000000002, latitude: -15.773 },
-    { longitude: -47.807000000000002, latitude: -15.777000000000001 },
-    { longitude: -47.803000000000004, latitude: -15.785 },
-    { longitude: -47.798000000000002, latitude: -15.786 },
-    { longitude: -47.794000000000004, latitude: -15.784000000000001 },
-    { longitude: -47.792000000000002, latitude: -15.786 },
-    { longitude: -47.791000000000004, latitude: -15.781000000000001 },
-    { longitude: -47.788000000000004, latitude: -15.778 },
-    { longitude: -47.785000000000004, latitude: -15.778 },
-    { longitude: -47.780999999999999, latitude: -15.781000000000001 },
-    { longitude: -47.779000000000003, latitude: -15.787000000000001 },
-    { longitude: -47.780000000000001, latitude: -15.800000000000001 },
-    { longitude: -47.783999999999999, latitude: -15.806000000000001 },
-    { longitude: -47.788000000000004, latitude: -15.806000000000001 },
-    { longitude: -47.792999999999999, latitude: -15.801 },
-    { longitude: -47.801000000000002, latitude: -15.801 },
-    { longitude: -47.803000000000004, latitude: -15.798 },
-    { longitude: -47.807000000000002, latitude: -15.798 },
-    { longitude: -47.810000000000002, latitude: -15.801 },
-    { longitude: -47.808999999999997, latitude: -15.806000000000001 },
-    { longitude: -47.810000000000002, latitude: -15.809000000000001 },
-    { longitude: -47.814, latitude: -15.81 },
-    { longitude: -47.814999999999998, latitude: -15.813000000000001 },
-    { longitude: -47.814999999999998, latitude: -15.817 },
-    { longitude: -47.813000000000002, latitude: -15.818 },
-    { longitude: -47.811999999999998, latitude: -15.822000000000001 },
-    { longitude: -47.814, latitude: -15.825000000000001 },
-    { longitude: -47.817999999999998, latitude: -15.825000000000001 },
-    { longitude: -47.820999999999998, latitude: -15.829000000000001 },
-    { longitude: -47.823, latitude: -15.829000000000001 },
-    { longitude: -47.820999999999998, latitude: -15.832000000000001 },
-    { longitude: -47.822000000000003, latitude: -15.835000000000001 },
-    { longitude: -47.825000000000003, latitude: -15.837 },
-    { longitude: -47.834000000000003, latitude: -15.835000000000001 },
-    { longitude: -47.838000000000001, latitude: -15.843 },
-    { longitude: -47.846000000000004, latitude: -15.844000000000001 },
-    { longitude: -47.852000000000004, latitude: -15.843 },
-    { longitude: -47.856000000000002, latitude: -15.847 },
-    { longitude: -47.853999999999999, latitude: -15.851000000000001 },
-    { longitude: -47.855000000000004, latitude: -15.855 },
-    { longitude: -47.856999999999999, latitude: -15.857000000000001 },
-    { longitude: -47.864000000000004, latitude: -15.856 },
-    { longitude: -47.870000000000005, latitude: -15.863 },
-    { longitude: -47.875999999999998, latitude: -15.862 },
-    { longitude: -47.878999999999998, latitude: -15.859 },
-    { longitude: -47.878999999999998, latitude: -15.856 },
-    { longitude: -47.872999999999998, latitude: -15.851000000000001 },
-    { longitude: -47.874000000000002, latitude: -15.846 },
-    { longitude: -47.864000000000004, latitude: -15.832000000000001 },
-    { longitude: -47.872, latitude: -15.834 },
-    { longitude: -47.875, latitude: -15.832000000000001 },
-    { longitude: -47.875999999999998, latitude: -15.829000000000001 },
-    { longitude: -47.880000000000003, latitude: -15.833 },
-    { longitude: -47.881999999999998, latitude: -15.839 },
-    { longitude: -47.890999999999998, latitude: -15.845000000000001 },
-    { longitude: -47.896999999999998, latitude: -15.845000000000001 },
-    { longitude: -47.905999999999999, latitude: -15.85 },
-    { longitude: -47.910000000000004, latitude: -15.85 },
-    { longitude: -47.919000000000004, latitude: -15.856 },
-    { longitude: -47.931000000000004, latitude: -15.856 },
-    { longitude: -47.936, latitude: -15.853 },
-    { longitude: -47.936, latitude: -15.849 },
-    { longitude: -47.941000000000003, latitude: -15.849 },
-    { longitude: -47.956000000000003, latitude: -15.844000000000001 },
-    { longitude: -47.957000000000001, latitude: -15.84 },
-    { longitude: -47.951999999999998, latitude: -15.808 },
-    { longitude: -47.948999999999998, latitude: -15.806000000000001 },
-    { longitude: -47.939999999999998, latitude: -15.807 },
-    { longitude: -47.920999999999999, latitude: -15.798 },
-    { longitude: -47.922000000000004, latitude: -15.797000000000001 },
-    { longitude: -47.919000000000004, latitude: -15.788 },
-    { longitude: -47.942, latitude: -15.780000000000001 },
-    { longitude: -47.941000000000003, latitude: -15.771000000000001 },
-    { longitude: -47.925000000000004, latitude: -15.746 },
-    { longitude: -47.928000000000004, latitude: -15.745000000000001 },
-    { longitude: -47.941000000000003, latitude: -15.752000000000001 },
-    { longitude: -47.957999999999998, latitude: -15.750999999999999 },
-    { longitude: -47.969999999999999, latitude: -15.758000000000001 },
-    { longitude: -47.972000000000001, latitude: -15.77 },
-    { longitude: -47.966999999999999, latitude: -15.775 },
-    { longitude: -47.966999999999999, latitude: -15.778 },
-    { longitude: -47.975000000000001, latitude: -15.786 },
-    { longitude: -47.978999999999999, latitude: -15.787000000000001 },
-    { longitude: -48.006, latitude: -15.758000000000001 },
-    { longitude: -48.017000000000003, latitude: -15.756 },
-    { longitude: -48.048999999999999, latitude: -15.758000000000001 },
-    { longitude: -48.056000000000004, latitude: -15.746 },
-    { longitude: -48.061, latitude: -15.743 },
-    { longitude: -48.073999999999998, latitude: -15.74 },
-];
-
-export const coordinatesSudoesteOctogonal = [
-    { longitude: -47.951999999999998, latitude: -15.809000000000001 },
-    { longitude: -47.948999999999998, latitude: -15.797000000000001 },
-    { longitude: -47.945, latitude: -15.796000000000001 },
-    { longitude: -47.941000000000003, latitude: -15.798 },
-    { longitude: -47.936999999999998, latitude: -15.796000000000001 },
-    { longitude: -47.938000000000002, latitude: -15.793000000000001 },
-    { longitude: -47.933, latitude: -15.778 },
-    { longitude: -47.931000000000004, latitude: -15.776 },
-    { longitude: -47.911999999999999, latitude: -15.781000000000001 },
-    { longitude: -47.910000000000004, latitude: -15.782999999999999 },
-    { longitude: -47.913000000000004, latitude: -15.794 },
-    { longitude: -47.910000000000004, latitude: -15.796000000000001 },
-    { longitude: -47.911000000000001, latitude: -15.802 },
-    { longitude: -47.938000000000002, latitude: -15.815 },
-    { longitude: -47.950000000000003, latitude: -15.814 },
-    { longitude: -47.951999999999998, latitude: -15.812000000000001 },
-    { longitude: -47.951999999999998, latitude: -15.809000000000001 },
-];
-
-export const coordinatesBrazlandia = [
-    { longitude: -48.245, latitude: -15.694 },
-    { longitude: -48.245, latitude: -15.684 },
-    { longitude: -48.238, latitude: -15.664 },
-    { longitude: -48.239, latitude: -15.664 },
-    { longitude: -48.238, latitude: -15.657 },
-    { longitude: -48.24, latitude: -15.655 },
-    { longitude: -48.239, latitude: -15.641 },
-    { longitude: -48.235, latitude: -15.638 },
-    { longitude: -48.231, latitude: -15.63 },
-    { longitude: -48.226, latitude: -15.625 },
-    { longitude: -48.224, latitude: -15.625 },
-    { longitude: -48.219, latitude: -15.619 },
-    { longitude: -48.204, latitude: -15.62 },
-    { longitude: -48.202, latitude: -15.616 },
-    { longitude: -48.205, latitude: -15.5 },
-    { longitude: -48.203, latitude: -15.497 },
-    { longitude: -48.035, latitude: -15.496 },
-    { longitude: -48.032, latitude: -15.498 },
-    { longitude: -48.03, latitude: -15.502 },
-    { longitude: -48.03, latitude: -15.507 },
-    { longitude: -48.036, latitude: -15.519 },
-    { longitude: -48.034, latitude: -15.527 },
-    { longitude: -48.03, latitude: -15.527 },
-    { longitude: -48.028, latitude: -15.529 },
-    { longitude: -48.028, latitude: -15.534 },
-    { longitude: -48.039, latitude: -15.544 },
-    { longitude: -48.04, latitude: -15.547 },
-    { longitude: -48.031, latitude: -15.553 },
-    { longitude: -48.029, latitude: -15.558 },
-    { longitude: -48.031, latitude: -15.564 },
-    { longitude: -48.021, latitude: -15.571 },
-    { longitude: -48.013, latitude: -15.58 },
-    { longitude: -48.013, latitude: -15.584 },
-    { longitude: -48.024, latitude: -15.594 },
-    { longitude: -48.04, latitude: -15.614 },
-    { longitude: -48.069, latitude: -15.628 },
-    { longitude: -48.073, latitude: -15.634 },
-    { longitude: -48.086, latitude: -15.679 },
-    { longitude: -48.082, latitude: -15.725 },
-    { longitude: -48.079, latitude: -15.729 },
-    { longitude: -48.07, latitude: -15.733 },
-    { longitude: -48.068, latitude: -15.736 },
-    { longitude: -48.069, latitude: -15.739 },
-    { longitude: -48.062, latitude: -15.752 },
-    { longitude: -48.062, latitude: -15.757 },
-    { longitude: -48.065, latitude: -15.762 },
-    { longitude: -48.075, latitude: -15.763 },
-    { longitude: -48.078, latitude: -15.761 },
-    { longitude: -48.079, latitude: -15.763 },
-    { longitude: -48.085, latitude: -15.764 },
-    { longitude: -48.09, latitude: -15.768 },
-    { longitude: -48.101, latitude: -15.77 },
-    { longitude: -48.111, latitude: -15.775 },
-    { longitude: -48.134, latitude: -15.775 },
-    { longitude: -48.137, latitude: -15.772 },
-    { longitude: -48.15, latitude: -15.773 },
-    { longitude: -48.16, latitude: -15.771 },
-    { longitude: -48.162, latitude: -15.769 },
-    { longitude: -48.163, latitude: -15.763 },
-    { longitude: -48.171, latitude: -15.756 },
-    { longitude: -48.176, latitude: -15.757 },
-    { longitude: -48.184, latitude: -15.752 },
-    { longitude: -48.206, latitude: -15.756 },
-    { longitude: -48.214, latitude: -15.755 },
-    { longitude: -48.218, latitude: -15.752 },
-    { longitude: -48.217, latitude: -15.748 },
-    { longitude: -48.22, latitude: -15.743 },
-    { longitude: -48.22, latitude: -15.739 },
-    { longitude: -48.217, latitude: -15.737 },
-    { longitude: -48.218, latitude: -15.729 },
-    { longitude: -48.215, latitude: -15.721 },
-    { longitude: -48.217, latitude: -15.719 },
-    { longitude: -48.217, latitude: -15.716 },
-    { longitude: -48.221, latitude: -15.713 },
-    { longitude: -48.233, latitude: -15.714 },
-    { longitude: -48.242, latitude: -15.711 },
-    { longitude: -48.245, latitude: -15.708 },
-    { longitude: -48.244, latitude: -15.701 },
-    { longitude: -48.245, latitude: -15.694 },
-];
-
-export const coordinatesCandangolandia = [
-    { longitude: -47.959, latitude: -15.879 },
-    { longitude: -47.963, latitude: -15.876 },
-    { longitude: -47.963, latitude: -15.874 },
-    { longitude: -47.957, latitude: -15.838 },
-    { longitude: -47.952, latitude: -15.837 },
-    { longitude: -47.94, latitude: -15.841 },
-    { longitude: -47.928, latitude: -15.842 },
-    { longitude: -47.927, latitude: -15.845 },
-    { longitude: -47.931, latitude: -15.863 },
-    { longitude: -47.95, latitude: -15.876 },
-    { longitude: -47.959, latitude: -15.879 },
-];
-
-export const coordinatesCruzeiro = [
-    { longitude: -47.95, latitude: -15.799 },
-    { longitude: -47.942, latitude: -15.775 },
-    { longitude: -47.94, latitude: -15.773 },
-    { longitude: -47.928, latitude: -15.776 },
-    { longitude: -47.925, latitude: -15.779 },
-    { longitude: -47.93, latitude: -15.793 },
-    { longitude: -47.929, latitude: -15.798 },
-    { longitude: -47.932, latitude: -15.802 },
-    { longitude: -47.938, latitude: -15.806 },
-    { longitude: -47.942, latitude: -15.806 },
-    { longitude: -47.948, latitude: -15.804 },
-    { longitude: -47.95, latitude: -15.802 },
-    { longitude: -47.95, latitude: -15.799 },
-];
-
-export const coordinatesEstruturalSCIA = [
-    { longitude: -48.011, latitude: -15.764 },
-    { longitude: -48.01, latitude: -15.753 },
-    { longitude: -48.009, latitude: -15.751 },
-    { longitude: -48.005, latitude: -15.75 },
-    { longitude: -48.001, latitude: -15.752 },
-    { longitude: -47.978, latitude: -15.777 },
-    { longitude: -47.977, latitude: -15.776 },
-    { longitude: -47.98, latitude: -15.773 },
-    { longitude: -47.98, latitude: -15.762 },
-    { longitude: -47.977, latitude: -15.754 },
-    { longitude: -47.964, latitude: -15.744 },
-    { longitude: -47.943, latitude: -15.744 },
-    { longitude: -47.929, latitude: -15.736 },
-    { longitude: -47.923, latitude: -15.738 },
-    { longitude: -47.921, latitude: -15.736 },
-    { longitude: -47.916, latitude: -15.736 },
-    { longitude: -47.913, latitude: -15.738 },
-    { longitude: -47.913, latitude: -15.741 },
-    { longitude: -47.931, latitude: -15.77 },
-    { longitude: -47.938, latitude: -15.788 },
-    { longitude: -47.954, latitude: -15.797 },
-    { longitude: -47.961, latitude: -15.799 },
-    { longitude: -47.971, latitude: -15.8 },
-    { longitude: -47.99, latitude: -15.791 },
-    { longitude: -48.005, latitude: -15.79 },
-    { longitude: -48.009, latitude: -15.784 },
-    { longitude: -48.011, latitude: -15.776 },
-    { longitude: -48.01, latitude: -15.772 },
-    { longitude: -48.011, latitude: -15.764 },
-];
-
-export const coordinatesFercal = [
-    { longitude: -47.933, latitude: -15.586 },
-    { longitude: -47.934, latitude: -15.582 },
-    { longitude: -47.931, latitude: -15.579 },
-    { longitude: -47.917, latitude: -15.581 },
-    { longitude: -47.897, latitude: -15.578 },
-    { longitude: -47.869, latitude: -15.554 },
-    { longitude: -47.845, latitude: -15.548 },
-    { longitude: -47.837, latitude: -15.541 },
-    { longitude: -47.817, latitude: -15.541 },
-    { longitude: -47.811, latitude: -15.545 },
-    { longitude: -47.808, latitude: -15.551 },
-    { longitude: -47.809, latitude: -15.562 },
-    { longitude: -47.843, latitude: -15.578 },
-    { longitude: -47.851, latitude: -15.595 },
-    { longitude: -47.851, latitude: -15.601 },
-    { longitude: -47.856, latitude: -15.614 },
-    { longitude: -47.852, latitude: -15.622 },
-    { longitude: -47.857, latitude: -15.626 },
-    { longitude: -47.861, latitude: -15.626 },
-    { longitude: -47.865, latitude: -15.63 },
-    { longitude: -47.873, latitude: -15.629 },
-    { longitude: -47.88, latitude: -15.623 },
-    { longitude: -47.879, latitude: -15.612 },
-    { longitude: -47.884, latitude: -15.609 },
-    { longitude: -47.884, latitude: -15.607 },
-    { longitude: -47.888, latitude: -15.604 },
-    { longitude: -47.891, latitude: -15.599 },
-    { longitude: -47.891, latitude: -15.597 },
-    { longitude: -47.896, latitude: -15.595 },
-    { longitude: -47.9, latitude: -15.598 },
-    { longitude: -47.904, latitude: -15.598 },
-    { longitude: -47.927, latitude: -15.591 },
-    { longitude: -47.933, latitude: -15.586 },
-];
-
-export const coordinatesItapoa = [
-    { longitude: -47.794, latitude: -15.729 },
-    { longitude: -47.794, latitude: -15.726 },
-    { longitude: -47.792, latitude: -15.724 },
-    { longitude: -47.781, latitude: -15.72 },
-    { longitude: -47.774, latitude: -15.72 },
-    { longitude: -47.762, latitude: -15.709 },
-    { longitude: -47.759, latitude: -15.71 },
-    { longitude: -47.757, latitude: -15.713 },
-    { longitude: -47.751, latitude: -15.715 },
-    { longitude: -47.747, latitude: -15.72 },
-    { longitude: -47.742, latitude: -15.721 },
-    { longitude: -47.74, latitude: -15.724 },
-    { longitude: -47.734, latitude: -15.722 },
-    { longitude: -47.728, latitude: -15.724 },
-    { longitude: -47.72, latitude: -15.721 },
-    { longitude: -47.716, latitude: -15.723 },
-    { longitude: -47.69, latitude: -15.722 },
-    { longitude: -47.68, latitude: -15.729 },
-    { longitude: -47.679, latitude: -15.728 },
-    { longitude: -47.673, latitude: -15.729 },
-    { longitude: -47.671, latitude: -15.731 },
-    { longitude: -47.671, latitude: -15.736 },
-    { longitude: -47.679, latitude: -15.74 },
-    { longitude: -47.689, latitude: -15.74 },
-    { longitude: -47.703, latitude: -15.747 },
-    { longitude: -47.729, latitude: -15.766 },
-    { longitude: -47.739, latitude: -15.767 },
-    { longitude: -47.757, latitude: -15.756 },
-    { longitude: -47.777, latitude: -15.762 },
-    { longitude: -47.782, latitude: -15.761 },
-    { longitude: -47.794, latitude: -15.729 },
-];
-
-export const coordinatesJardimBotanico = [
-    { longitude: -47.826, latitude: -15.886 },
-    { longitude: -47.827, latitude: -15.871 },
-    { longitude: -47.819, latitude: -15.847 },
-    { longitude: -47.808, latitude: -15.833 },
-    { longitude: -47.803, latitude: -15.829 },
-    { longitude: -47.796, latitude: -15.83 },
-    { longitude: -47.791, latitude: -15.837 },
-    { longitude: -47.789, latitude: -15.835 },
-    { longitude: -47.779, latitude: -15.834 },
-    { longitude: -47.772, latitude: -15.83 },
-    { longitude: -47.753, latitude: -15.84 },
-    { longitude: -47.742, latitude: -15.851 },
-    { longitude: -47.741, latitude: -15.854 },
-    { longitude: -47.743, latitude: -15.857 },
-    { longitude: -47.739, latitude: -15.861 },
-    { longitude: -47.738, latitude: -15.87 },
-    { longitude: -47.74, latitude: -15.873 },
-    { longitude: -47.748, latitude: -15.874 },
-    { longitude: -47.748, latitude: -15.89 },
-    { longitude: -47.752, latitude: -15.894 },
-    { longitude: -47.76, latitude: -15.895 },
-    { longitude: -47.767, latitude: -15.888 },
-    { longitude: -47.775, latitude: -15.89 },
-    { longitude: -47.781, latitude: -15.885 },
-    { longitude: -47.787, latitude: -15.883 },
-    { longitude: -47.791, latitude: -15.879 },
-    { longitude: -47.792, latitude: -15.885 },
-    { longitude: -47.79, latitude: -15.893 },
-    { longitude: -47.795, latitude: -15.895 },
-    { longitude: -47.817, latitude: -15.889 },
-    { longitude: -47.824, latitude: -15.889 },
-    { longitude: -47.826, latitude: -15.886 },
-];
-
-export const coordinatesLagoNorte = [
-    { longitude: -47.911, latitude: -15.732 },
-    { longitude: -47.913, latitude: -15.731 },
-    { longitude: -47.914, latitude: -15.728 },
-    { longitude: -47.902, latitude: -15.709 },
-    { longitude: -47.894, latitude: -15.703 },
-    { longitude: -47.888, latitude: -15.704 },
-    { longitude: -47.889, latitude: -15.701 },
-    { longitude: -47.887, latitude: -15.698 },
-    { longitude: -47.859, latitude: -15.684 },
-    { longitude: -47.85, latitude: -15.689 },
-    { longitude: -47.836, latitude: -15.689 },
-    { longitude: -47.832, latitude: -15.691 },
-    { longitude: -47.791, latitude: -15.719 },
-    { longitude: -47.786, latitude: -15.726 },
-    { longitude: -47.774, latitude: -15.76 },
-    { longitude: -47.778, latitude: -15.763 },
-    { longitude: -47.787, latitude: -15.766 },
-    { longitude: -47.805, latitude: -15.763 },
-    { longitude: -47.805, latitude: -15.767 },
-    { longitude: -47.809, latitude: -15.771 },
-    { longitude: -47.802, latitude: -15.782 },
-    { longitude: -47.799, latitude: -15.778 },
-    { longitude: -47.796, latitude: -15.777 },
-    { longitude: -47.785, latitude: -15.778 },
-    { longitude: -47.783, latitude: -15.78 },
-    { longitude: -47.783, latitude: -15.785 },
-    { longitude: -47.785, latitude: -15.792 },
-    { longitude: -47.787, latitude: -15.794 },
-    { longitude: -47.798, latitude: -15.795 },
-    { longitude: -47.809, latitude: -15.791 },
-    { longitude: -47.814, latitude: -15.782 },
-    { longitude: -47.82, latitude: -15.778 },
-    { longitude: -47.823, latitude: -15.762 },
-    { longitude: -47.827, latitude: -15.775 },
-    { longitude: -47.831, latitude: -15.778 },
-    { longitude: -47.836, latitude: -15.777 },
-    { longitude: -47.839, latitude: -15.773 },
-    { longitude: -47.839, latitude: -15.768 },
-    { longitude: -47.843, latitude: -15.767 },
-    { longitude: -47.844, latitude: -15.764 },
-    { longitude: -47.847, latitude: -15.764 },
-    { longitude: -47.85, latitude: -15.757 },
-    { longitude: -47.853, latitude: -15.756 },
-    { longitude: -47.854, latitude: -15.754 },
-    { longitude: -47.865, latitude: -15.751 },
-    { longitude: -47.867, latitude: -15.747 },
-    { longitude: -47.876, latitude: -15.744 },
-    { longitude: -47.882, latitude: -15.734 },
-    { longitude: -47.897, latitude: -15.734 },
-    { longitude: -47.898, latitude: -15.732 },
-    { longitude: -47.904, latitude: -15.731 },
-    { longitude: -47.908, latitude: -15.733 },
-    { longitude: -47.911, latitude: -15.732 },
-];
-
-export const coordinatesLagoSul = [
-    { longitude: -47.987, latitude: -15.954 },
-    { longitude: -47.988, latitude: -15.951 },
-    { longitude: -47.984, latitude: -15.948 },
-    { longitude: -47.973, latitude: -15.948 },
-    { longitude: -47.966, latitude: -15.946 },
-    { longitude: -47.966, latitude: -15.947 },
-    { longitude: -47.964, latitude: -15.944 },
-    { longitude: -47.958, latitude: -15.941 },
-    { longitude: -47.949, latitude: -15.94 },
-    { longitude: -47.943, latitude: -15.937 },
-    { longitude: -47.935, latitude: -15.93 },
-    { longitude: -47.933, latitude: -15.918 },
-    { longitude: -47.93, latitude: -15.911 },
-    { longitude: -47.923, latitude: -15.906 },
-    { longitude: -47.915, latitude: -15.904 },
-    { longitude: -47.914, latitude: -15.897 },
-    { longitude: -47.91, latitude: -15.891 },
-    { longitude: -47.91, latitude: -15.886 },
-    { longitude: -47.921, latitude: -15.888 },
-    { longitude: -47.94, latitude: -15.9 },
-    { longitude: -47.95, latitude: -15.897 },
-    { longitude: -47.955, latitude: -15.887 },
-    { longitude: -47.96, latitude: -15.883 },
-    { longitude: -47.959, latitude: -15.877 },
-    { longitude: -47.941, latitude: -15.868 },
-    { longitude: -47.935, latitude: -15.848 },
-    { longitude: -47.933, latitude: -15.847 },
-    { longitude: -47.922, latitude: -15.848 },
-    { longitude: -47.916, latitude: -15.844 },
-    { longitude: -47.903, latitude: -15.84 },
-    { longitude: -47.9, latitude: -15.837 },
-    { longitude: -47.893, latitude: -15.837 },
-    { longitude: -47.888, latitude: -15.833 },
-    { longitude: -47.887, latitude: -15.826 },
-    { longitude: -47.879, latitude: -15.821 },
-    { longitude: -47.874, latitude: -15.82 },
-    { longitude: -47.871, latitude: -15.82 },
-    { longitude: -47.867, latitude: -15.824 },
-    { longitude: -47.855, latitude: -15.823 },
-    { longitude: -47.851, latitude: -15.826 },
-    { longitude: -47.851, latitude: -15.831 },
-    { longitude: -47.857, latitude: -15.837 },
-    { longitude: -47.853, latitude: -15.835 },
-    { longitude: -47.844, latitude: -15.835 },
-    { longitude: -47.841, latitude: -15.83 },
-    { longitude: -47.836, latitude: -15.827 },
-    { longitude: -47.831, latitude: -15.827 },
-    { longitude: -47.825, latitude: -15.821 },
-    { longitude: -47.822, latitude: -15.812 },
-    { longitude: -47.823, latitude: -15.807 },
-    { longitude: -47.821, latitude: -15.806 },
-    { longitude: -47.814, latitude: -15.792 },
-    { longitude: -47.811, latitude: -15.789 },
-    { longitude: -47.803, latitude: -15.788 },
-    { longitude: -47.797, latitude: -15.792 },
-    { longitude: -47.791, latitude: -15.791 },
-    { longitude: -47.786, latitude: -15.793 },
-    { longitude: -47.785, latitude: -15.798 },
-    { longitude: -47.782, latitude: -15.8 },
-    { longitude: -47.783, latitude: -15.807 },
-    { longitude: -47.792, latitude: -15.826 },
-    { longitude: -47.812, latitude: -15.851 },
-    { longitude: -47.819, latitude: -15.872 },
-    { longitude: -47.819, latitude: -15.882 },
-    { longitude: -47.815, latitude: -15.898 },
-    { longitude: -47.816, latitude: -15.906 },
-    { longitude: -47.829, latitude: -15.931 },
-    { longitude: -47.834, latitude: -15.933 },
-    { longitude: -47.875, latitude: -15.967 },
-    { longitude: -47.933, latitude: -15.988 },
-    { longitude: -47.947, latitude: -15.987 },
-    { longitude: -47.958, latitude: -15.983 },
-    { longitude: -47.96, latitude: -15.981 },
-    { longitude: -47.961, latitude: -15.972 },
-    { longitude: -47.968, latitude: -15.972 },
-    { longitude: -47.972, latitude: -15.968 },
-    { longitude: -47.972, latitude: -15.965 },
-    { longitude: -47.978, latitude: -15.965 },
-    { longitude: -47.982, latitude: -15.963 },
-    { longitude: -47.987, latitude: -15.954 },
-];
-
-export const coordinatesNucleoBandeirante = [
-    { longitude: -48.001, latitude: -15.873 },
-    { longitude: -47.984, latitude: -15.872 },
-    { longitude: -47.978, latitude: -15.863 },
-    { longitude: -47.967, latitude: -15.861 },
-    { longitude: -47.965, latitude: -15.851 },
-    { longitude: -47.96, latitude: -15.85 },
-    { longitude: -47.958, latitude: -15.845 },
-    { longitude: -47.954, latitude: -15.844 },
-    { longitude: -47.951, latitude: -15.846 },
-    { longitude: -47.955, latitude: -15.878 },
-    { longitude: -47.958, latitude: -15.881 },
-    { longitude: -47.966, latitude: -15.88 },
-    { longitude: -47.968, latitude: -15.887 },
-    { longitude: -47.979, latitude: -15.887 },
-    { longitude: -47.982, latitude: -15.883 },
-    { longitude: -48.001, latitude: -15.884 },
-    { longitude: -48.004, latitude: -15.882 },
-    { longitude: -48.005, latitude: -15.877 },
-    { longitude: -48.001, latitude: -15.873 },
-];
-
-export const coordinatesParanoa = [
-    { longitude: -47.813, latitude: -15.764 },
-    { longitude: -47.815, latitude: -15.763 },
-    { longitude: -47.817, latitude: -15.758 },
-    { longitude: -47.812, latitude: -15.753 },
-    { longitude: -47.788, latitude: -15.758 },
-    { longitude: -47.762, latitude: -15.748 },
-    { longitude: -47.755, latitude: -15.748 },
-    { longitude: -47.737, latitude: -15.759 },
-    { longitude: -47.733, latitude: -15.759 },
-    { longitude: -47.707, latitude: -15.74 },
-    { longitude: -47.691, latitude: -15.732 },
-    { longitude: -47.681, latitude: -15.732 },
-    { longitude: -47.679, latitude: -15.73 },
-    { longitude: -47.674, latitude: -15.729 },
-    { longitude: -47.668, latitude: -15.724 },
-    { longitude: -47.66, latitude: -15.723 },
-    { longitude: -47.657, latitude: -15.725 },
-    { longitude: -47.657, latitude: -15.728 },
-    { longitude: -47.661, latitude: -15.748 },
-    { longitude: -47.638, latitude: -15.788 },
-    { longitude: -47.618, latitude: -15.858 },
-    { longitude: -47.608, latitude: -15.855 },
-    { longitude: -47.602, latitude: -15.855 },
-    { longitude: -47.559, latitude: -15.868 },
-    { longitude: -47.555, latitude: -15.868 },
-    { longitude: -47.522, latitude: -15.85 },
-    { longitude: -47.515, latitude: -15.849 },
-    { longitude: -47.467, latitude: -15.859 },
-    { longitude: -47.457, latitude: -15.854 },
-    { longitude: -47.451, latitude: -15.854 },
-    { longitude: -47.423, latitude: -15.86 },
-    { longitude: -47.391, latitude: -15.854 },
-    { longitude: -47.389, latitude: -15.843 },
-    { longitude: -47.387, latitude: -15.842 },
-    { longitude: -47.368, latitude: -15.842 },
-    { longitude: -47.365, latitude: -15.845 },
-    { longitude: -47.364, latitude: -15.855 },
-    { longitude: -47.366, latitude: -15.856 },
-    { longitude: -47.365, latitude: -15.857 },
-    { longitude: -47.368, latitude: -15.876 },
-    { longitude: -47.37, latitude: -15.877 },
-    { longitude: -47.37, latitude: -15.88 },
-    { longitude: -47.372, latitude: -15.881 },
-    { longitude: -47.374, latitude: -15.886 },
-    { longitude: -47.372, latitude: -15.894 },
-    { longitude: -47.368, latitude: -15.899 },
-    { longitude: -47.369, latitude: -15.904 },
-    { longitude: -47.364, latitude: -15.905 },
-    { longitude: -47.361, latitude: -15.909 },
-    { longitude: -47.363, latitude: -15.916 },
-    { longitude: -47.358, latitude: -15.92 },
-    { longitude: -47.357, latitude: -15.927 },
-    { longitude: -47.359, latitude: -15.934 },
-    { longitude: -47.363, latitude: -15.937 },
-    { longitude: -47.366, latitude: -15.937 },
-    { longitude: -47.366, latitude: -15.944 },
-    { longitude: -47.364, latitude: -15.948 },
-    { longitude: -47.367, latitude: -15.951 },
-    { longitude: -47.37, latitude: -15.963 },
-    { longitude: -47.372, latitude: -15.964 },
-    { longitude: -47.371, latitude: -15.967 },
-    { longitude: -47.374, latitude: -15.975 },
-    { longitude: -47.37, latitude: -15.978 },
-    { longitude: -47.37, latitude: -15.983 },
-    { longitude: -47.372, latitude: -15.984 },
-    { longitude: -47.37, latitude: -15.991 },
-    { longitude: -47.367, latitude: -15.992 },
-    { longitude: -47.366, latitude: -15.997 },
-    { longitude: -47.362, latitude: -15.998 },
-    { longitude: -47.361, latitude: -16.002 },
-    { longitude: -47.356, latitude: -16.003 },
-    { longitude: -47.347, latitude: -16.002 },
-    { longitude: -47.342, latitude: -16.008 },
-    { longitude: -47.338, latitude: -16.008 },
-    { longitude: -47.335, latitude: -16.011 },
-    { longitude: -47.332, latitude: -16.018 },
-    { longitude: -47.328, latitude: -16.019 },
-    { longitude: -47.326, latitude: -16.023 },
-    { longitude: -47.319, latitude: -16.023 },
-    { longitude: -47.318, latitude: -16.028 },
-    { longitude: -47.315, latitude: -16.031 },
-    { longitude: -47.309, latitude: -16.032 },
-    { longitude: -47.306, latitude: -16.035 },
-    { longitude: -47.306, latitude: -16.039 },
-    { longitude: -47.304, latitude: -16.042 },
-    { longitude: -47.305, latitude: -16.053 },
-    { longitude: -47.62, latitude: -16.054 },
-    { longitude: -47.622, latitude: -16.049 },
-    { longitude: -47.618, latitude: -16.012 },
-    { longitude: -47.592, latitude: -15.972 },
-    { longitude: -47.592, latitude: -15.97 },
-    { longitude: -47.61, latitude: -15.941 },
-    { longitude: -47.616, latitude: -15.942 },
-    { longitude: -47.627, latitude: -15.95 },
-    { longitude: -47.627, latitude: -15.955 },
-    { longitude: -47.629, latitude: -15.958 },
-    { longitude: -47.648, latitude: -15.973 },
-    { longitude: -47.654, latitude: -15.973 },
-    { longitude: -47.664, latitude: -15.966 },
-    { longitude: -47.668, latitude: -15.961 },
-    { longitude: -47.667, latitude: -15.957 },
-    { longitude: -47.67, latitude: -15.955 },
-    { longitude: -47.673, latitude: -15.947 },
-    { longitude: -47.673, latitude: -15.935 },
-    { longitude: -47.68, latitude: -15.933 },
-    { longitude: -47.681, latitude: -15.926 },
-    { longitude: -47.685, latitude: -15.925 },
-    { longitude: -47.687, latitude: -15.918 },
-    { longitude: -47.692, latitude: -15.914 },
-    { longitude: -47.692, latitude: -15.912 },
-    { longitude: -47.699, latitude: -15.909 },
-    { longitude: -47.702, latitude: -15.906 },
-    { longitude: -47.703, latitude: -15.902 },
-    { longitude: -47.713, latitude: -15.894 },
-    { longitude: -47.716, latitude: -15.889 },
-    { longitude: -47.716, latitude: -15.885 },
-    { longitude: -47.714, latitude: -15.882 },
-    { longitude: -47.723, latitude: -15.881 },
-    { longitude: -47.743, latitude: -15.874 },
-    { longitude: -47.747, latitude: -15.868 },
-    { longitude: -47.747, latitude: -15.865 },
-    { longitude: -47.751, latitude: -15.861 },
-    { longitude: -47.752, latitude: -15.853 },
-    { longitude: -47.758, latitude: -15.847 },
-    { longitude: -47.772, latitude: -15.839 },
-    { longitude: -47.78, latitude: -15.843 },
-    { longitude: -47.785, latitude: -15.843 },
-    { longitude: -47.79, latitude: -15.848 },
-    { longitude: -47.794, latitude: -15.848 },
-    { longitude: -47.8, latitude: -15.837 },
-    { longitude: -47.803, latitude: -15.839 },
-    { longitude: -47.806, latitude: -15.839 },
-    { longitude: -47.808, latitude: -15.837 },
-    { longitude: -47.808, latitude: -15.833 },
-    { longitude: -47.8, latitude: -15.824 },
-    { longitude: -47.787, latitude: -15.796 },
-    { longitude: -47.787, latitude: -15.786 },
-    { longitude: -47.795, latitude: -15.785 },
-    { longitude: -47.798, latitude: -15.79 },
-    { longitude: -47.803, latitude: -15.792 },
-    { longitude: -47.806, latitude: -15.791 },
-    { longitude: -47.817, latitude: -15.774 },
-    { longitude: -47.817, latitude: -15.769 },
-    { longitude: -47.813, latitude: -15.764 },
-];
-
-export const coordinatesParkWay = [
-    { longitude: -48.023, latitude: -15.963 },
-    { longitude: -48.005, latitude: -15.947 },
-    { longitude: -48.006, latitude: -15.945 },
-    { longitude: -48.003, latitude: -15.939 },
-    { longitude: -47.996, latitude: -15.932 },
-    { longitude: -47.988, latitude: -15.927 },
-    { longitude: -47.986, latitude: -15.918 },
-    { longitude: -47.993, latitude: -15.897 },
-    { longitude: -47.993, latitude: -15.885 },
-    { longitude: -47.989, latitude: -15.88 },
-    { longitude: -47.982, latitude: -15.877 },
-    { longitude: -47.976, latitude: -15.877 },
-    { longitude: -47.975, latitude: -15.879 },
-    { longitude: -47.974, latitude: -15.876 },
-    { longitude: -47.968, latitude: -15.871 },
-    { longitude: -47.958, latitude: -15.871 },
-    { longitude: -47.936, latitude: -15.857 },
-    { longitude: -47.933, latitude: -15.857 },
-    { longitude: -47.931, latitude: -15.86 },
-    { longitude: -47.933, latitude: -15.867 },
-    { longitude: -47.932, latitude: -15.872 },
-    { longitude: -47.95, latitude: -15.881 },
-    { longitude: -47.944, latitude: -15.891 },
-    { longitude: -47.936, latitude: -15.889 },
-    { longitude: -47.935, latitude: -15.887 },
-    { longitude: -47.933, latitude: -15.887 },
-    { longitude: -47.924, latitude: -15.88 },
-    { longitude: -47.912, latitude: -15.879 },
-    { longitude: -47.91, latitude: -15.877 },
-    { longitude: -47.903, latitude: -15.878 },
-    { longitude: -47.902, latitude: -15.893 },
-    { longitude: -47.907, latitude: -15.901 },
-    { longitude: -47.908, latitude: -15.909 },
-    { longitude: -47.911, latitude: -15.912 },
-    { longitude: -47.918, latitude: -15.913 },
-    { longitude: -47.924, latitude: -15.917 },
-    { longitude: -47.927, latitude: -15.933 },
-    { longitude: -47.932, latitude: -15.939 },
-    { longitude: -47.94, latitude: -15.945 },
-    { longitude: -47.946, latitude: -15.948 },
-    { longitude: -47.956, latitude: -15.949 },
-    { longitude: -47.964, latitude: -15.955 },
-    { longitude: -47.966, latitude: -15.954 },
-    { longitude: -47.971, latitude: -15.956 },
-    { longitude: -47.977, latitude: -15.956 },
-    { longitude: -47.969, latitude: -15.957 },
-    { longitude: -47.966, latitude: -15.959 },
-    { longitude: -47.964, latitude: -15.964 },
-    { longitude: -47.959, latitude: -15.963 },
-    { longitude: -47.953, latitude: -15.967 },
-    { longitude: -47.952, latitude: -15.981 },
-    { longitude: -47.954, latitude: -15.983 },
-    { longitude: -47.974, latitude: -15.978 },
-    { longitude: -47.996, latitude: -15.982 },
-    { longitude: -48.009, latitude: -15.981 },
-    { longitude: -48.018, latitude: -15.976 },
-    { longitude: -48.024, latitude: -15.969 },
-    { longitude: -48.025, latitude: -15.965 },
-    { longitude: -48.023, latitude: -15.963 },
-];
-
-export const coordinatesPlanaltina = [
-    { longitude: -47.778, latitude: -15.603 },
-    { longitude: -47.779, latitude: -15.599 },
-    { longitude: -47.77, latitude: -15.584 },
-    { longitude: -47.769, latitude: -15.579 },
-    { longitude: -47.765, latitude: -15.576 },
-    { longitude: -47.768, latitude: -15.572 },
-    { longitude: -47.767, latitude: -15.569 },
-    { longitude: -47.771, latitude: -15.566 },
-    { longitude: -47.771, latitude: -15.558 },
-    { longitude: -47.768, latitude: -15.555 },
-    { longitude: -47.76, latitude: -15.552 },
-    { longitude: -47.761, latitude: -15.546 },
-    { longitude: -47.759, latitude: -15.542 },
-    { longitude: -47.755, latitude: -15.539 },
-    { longitude: -47.757, latitude: -15.535 },
-    { longitude: -47.762, latitude: -15.534 },
-    { longitude: -47.767, latitude: -15.53 },
-    { longitude: -47.768, latitude: -15.525 },
-    { longitude: -47.772, latitude: -15.523 },
-    { longitude: -47.772, latitude: -15.52 },
-    { longitude: -47.778, latitude: -15.517 },
-    { longitude: -47.778, latitude: -15.513 },
-    { longitude: -47.774, latitude: -15.506 },
-    { longitude: -47.774, latitude: -15.498 },
-    { longitude: -47.771, latitude: -15.496 },
-    { longitude: -47.617, latitude: -15.496 },
-    { longitude: -47.614, latitude: -15.498 },
-    { longitude: -47.613, latitude: -15.501 },
-    { longitude: -47.572, latitude: -15.51 },
-    { longitude: -47.566, latitude: -15.498 },
-    { longitude: -47.563, latitude: -15.496 },
-    { longitude: -47.417, latitude: -15.496 },
-    { longitude: -47.414, latitude: -15.497 },
-    { longitude: -47.413, latitude: -15.5 },
-    { longitude: -47.413, latitude: -15.544 },
-    { longitude: -47.4, latitude: -15.544 },
-    { longitude: -47.399, latitude: -15.547 },
-    { longitude: -47.396, latitude: -15.548 },
-    { longitude: -47.395, latitude: -15.554 },
-    { longitude: -47.392, latitude: -15.554 },
-    { longitude: -47.39, latitude: -15.556 },
-    { longitude: -47.39, latitude: -15.559 },
-    { longitude: -47.388, latitude: -15.56 },
-    { longitude: -47.387, latitude: -15.563 },
-    { longitude: -47.379, latitude: -15.563 },
-    { longitude: -47.377, latitude: -15.565 },
-    { longitude: -47.374, latitude: -15.573 },
-    { longitude: -47.368, latitude: -15.575 },
-    { longitude: -47.36, latitude: -15.574 },
-    { longitude: -47.359, latitude: -15.576 },
-    { longitude: -47.346, latitude: -15.575 },
-    { longitude: -47.334, latitude: -15.583 },
-    { longitude: -47.319, latitude: -15.583 },
-    { longitude: -47.313, latitude: -15.589 },
-    { longitude: -47.312, latitude: -15.599 },
-    { longitude: -47.316, latitude: -15.602 },
-    { longitude: -47.319, latitude: -15.602 },
-    { longitude: -47.319, latitude: -15.606 },
-    { longitude: -47.324, latitude: -15.615 },
-    { longitude: -47.324, latitude: -15.62 },
-    { longitude: -47.329, latitude: -15.632 },
-    { longitude: -47.325, latitude: -15.639 },
-    { longitude: -47.325, latitude: -15.644 },
-    { longitude: -47.323, latitude: -15.646 },
-    { longitude: -47.324, latitude: -15.659 },
-    { longitude: -47.322, latitude: -15.671 },
-    { longitude: -47.323, latitude: -15.679 },
-    { longitude: -47.325, latitude: -15.681 },
-    { longitude: -47.324, latitude: -15.684 },
-    { longitude: -47.323, latitude: -15.683 },
-    { longitude: -47.317, latitude: -15.684 },
-    { longitude: -47.313, latitude: -15.691 },
-    { longitude: -47.309, latitude: -15.694 },
-    { longitude: -47.31, latitude: -15.713 },
-    { longitude: -47.315, latitude: -15.722 },
-    { longitude: -47.313, latitude: -15.733 },
-    { longitude: -47.309, latitude: -15.738 },
-    { longitude: -47.31, latitude: -15.748 },
-    { longitude: -47.315, latitude: -15.753 },
-    { longitude: -47.321, latitude: -15.765 },
-    { longitude: -47.323, latitude: -15.765 },
-    { longitude: -47.327, latitude: -15.781 },
-    { longitude: -47.332, latitude: -15.786 },
-    { longitude: -47.333, latitude: -15.789 },
-    { longitude: -47.335, latitude: -15.789 },
-    { longitude: -47.336, latitude: -15.797 },
-    { longitude: -47.342, latitude: -15.803 },
-    { longitude: -47.347, latitude: -15.812 },
-    { longitude: -47.351, latitude: -15.824 },
-    { longitude: -47.353, latitude: -15.827 },
-    { longitude: -47.358, latitude: -15.828 },
-    { longitude: -47.36, latitude: -15.831 },
-    { longitude: -47.362, latitude: -15.831 },
-    { longitude: -47.362, latitude: -15.839 },
-    { longitude: -47.366, latitude: -15.847 },
-    { longitude: -47.369, latitude: -15.85 },
-    { longitude: -47.382, latitude: -15.851 },
-    { longitude: -47.383, latitude: -15.858 },
-    { longitude: -47.386, latitude: -15.861 },
-    { longitude: -47.422, latitude: -15.868 },
-    { longitude: -47.452, latitude: -15.862 },
-    { longitude: -47.464, latitude: -15.867 },
-    { longitude: -47.468, latitude: -15.867 },
-    { longitude: -47.519, latitude: -15.857 },
-    { longitude: -47.553, latitude: -15.876 },
-    { longitude: -47.561, latitude: -15.876 },
-    { longitude: -47.604, latitude: -15.863 },
-    { longitude: -47.623, latitude: -15.867 },
-    { longitude: -47.625, latitude: -15.865 },
-    { longitude: -47.646, latitude: -15.79 },
-    { longitude: -47.669, latitude: -15.749 },
-    { longitude: -47.666, latitude: -15.733 },
-    { longitude: -47.672, latitude: -15.738 },
-    { longitude: -47.678, latitude: -15.738 },
-    { longitude: -47.68, latitude: -15.736 },
-    { longitude: -47.679, latitude: -15.73 },
-    { longitude: -47.681, latitude: -15.727 },
-    { longitude: -47.68, latitude: -15.724 },
-    { longitude: -47.681, latitude: -15.721 },
-    { longitude: -47.679, latitude: -15.711 },
-    { longitude: -47.681, latitude: -15.709 },
-    { longitude: -47.684, latitude: -15.709 },
-    { longitude: -47.686, latitude: -15.707 },
-    { longitude: -47.692, latitude: -15.707 },
-    { longitude: -47.699, latitude: -15.703 },
-    { longitude: -47.712, latitude: -15.704 },
-    { longitude: -47.721, latitude: -15.701 },
-    { longitude: -47.734, latitude: -15.702 },
-    { longitude: -47.744, latitude: -15.692 },
-    { longitude: -47.75, latitude: -15.689 },
-    { longitude: -47.751, latitude: -15.68 },
-    { longitude: -47.747, latitude: -15.673 },
-    { longitude: -47.755, latitude: -15.633 },
-    { longitude: -47.778, latitude: -15.603 },
-];
-
-export const coordinatesRecantoDasEmas = [
-    { longitude: -48.264, latitude: -15.967 },
-    { longitude: -48.261, latitude: -15.959 },
-    { longitude: -48.256, latitude: -15.953 },
-    { longitude: -48.258, latitude: -15.946 },
-    { longitude: -48.255, latitude: -15.939 },
-    { longitude: -48.239, latitude: -15.938 },
-    { longitude: -48.229, latitude: -15.933 },
-    { longitude: -48.224, latitude: -15.933 },
-    { longitude: -48.202, latitude: -15.92 },
-    { longitude: -48.191, latitude: -15.916 },
-    { longitude: -48.185, latitude: -15.916 },
-    { longitude: -48.183, latitude: -15.918 },
-    { longitude: -48.18, latitude: -15.915 },
-    { longitude: -48.173, latitude: -15.914 },
-    { longitude: -48.169, latitude: -15.917 },
-    { longitude: -48.167, latitude: -15.912 },
-    { longitude: -48.163, latitude: -15.909 },
-    { longitude: -48.061, latitude: -15.878 },
-    { longitude: -48.058, latitude: -15.879 },
-    { longitude: -48.056, latitude: -15.882 },
-    { longitude: -48.039, latitude: -15.936 },
-    { longitude: -48.034, latitude: -15.944 },
-    { longitude: -48.036, latitude: -15.95 },
-    { longitude: -48.043, latitude: -15.956 },
-    { longitude: -48.055, latitude: -15.957 },
-    { longitude: -48.077, latitude: -15.943 },
-    { longitude: -48.081, latitude: -15.942 },
-    { longitude: -48.082, latitude: -15.931 },
-    { longitude: -48.088, latitude: -15.931 },
-    { longitude: -48.092, latitude: -15.936 },
-    { longitude: -48.096, latitude: -15.937 },
-    { longitude: -48.099, latitude: -15.94 },
-    { longitude: -48.102, latitude: -15.947 },
-    { longitude: -48.11, latitude: -15.955 },
-    { longitude: -48.113, latitude: -15.955 },
-    { longitude: -48.118, latitude: -15.958 },
-    { longitude: -48.119, latitude: -15.965 },
-    { longitude: -48.114, latitude: -15.972 },
-    { longitude: -48.114, latitude: -15.977 },
-    { longitude: -48.119, latitude: -15.983 },
-    { longitude: -48.125, latitude: -15.982 },
-    { longitude: -48.137, latitude: -15.958 },
-    { longitude: -48.145, latitude: -15.955 },
-    { longitude: -48.148, latitude: -15.949 },
-    { longitude: -48.155, latitude: -15.951 },
-    { longitude: -48.157, latitude: -15.949 },
-    { longitude: -48.159, latitude: -15.941 },
-    { longitude: -48.161, latitude: -15.946 },
-    { longitude: -48.163, latitude: -15.947 },
-    { longitude: -48.168, latitude: -15.945 },
-    { longitude: -48.171, latitude: -15.951 },
-    { longitude: -48.183, latitude: -15.951 },
-    { longitude: -48.185, latitude: -15.959 },
-    { longitude: -48.192, latitude: -15.968 },
-    { longitude: -48.201, latitude: -15.967 },
-    { longitude: -48.21, latitude: -15.969 },
-    { longitude: -48.221, latitude: -15.975 },
-    { longitude: -48.241, latitude: -15.976 },
-    { longitude: -48.248, latitude: -15.98 },
-    { longitude: -48.262, latitude: -15.982 },
-    { longitude: -48.264, latitude: -15.975 },
-    { longitude: -48.263, latitude: -15.972 },
-    { longitude: -48.265, latitude: -15.969 },
-    { longitude: -48.264, latitude: -15.967 },
-];
-
-export const coordinatesRiachoFundoI = [
-    { longitude: -48.056, latitude: -15.877 },
-    { longitude: -48.046, latitude: -15.87 },
-    { longitude: -48.013, latitude: -15.874 },
-    { longitude: -47.999, latitude: -15.873 },
-    { longitude: -47.997, latitude: -15.876 },
-    { longitude: -47.978, latitude: -15.875 },
-    { longitude: -47.975, latitude: -15.879 },
-    { longitude: -47.976, latitude: -15.884 },
-    { longitude: -47.985, latitude: -15.887 },
-    { longitude: -47.985, latitude: -15.895 },
-    { longitude: -47.983, latitude: -15.898 },
-    { longitude: -47.984, latitude: -15.901 },
-    { longitude: -47.982, latitude: -15.903 },
-    { longitude: -47.982, latitude: -15.907 },
-    { longitude: -47.978, latitude: -15.916 },
-    { longitude: -47.978, latitude: -15.924 },
-    { longitude: -47.98, latitude: -15.925 },
-    { longitude: -47.979, latitude: -15.928 },
-    { longitude: -47.981, latitude: -15.932 },
-    { longitude: -47.993, latitude: -15.94 },
-    { longitude: -47.999, latitude: -15.949 },
-    { longitude: -48.006, latitude: -15.948 },
-    { longitude: -48.01, latitude: -15.941 },
-    { longitude: -48.016, latitude: -15.935 },
-    { longitude: -48.016, latitude: -15.933 },
-    { longitude: -48.013, latitude: -15.93 },
-    { longitude: -48.009, latitude: -15.929 },
-    { longitude: -48.009, latitude: -15.927 },
-    { longitude: -48.003, latitude: -15.92 },
-    { longitude: -48.006, latitude: -15.914 },
-    { longitude: -48.015, latitude: -15.909 },
-    { longitude: -48.025, latitude: -15.9 },
-    { longitude: -48.028, latitude: -15.9 },
-    { longitude: -48.03, latitude: -15.903 },
-    { longitude: -48.041, latitude: -15.903 },
-    { longitude: -48.045, latitude: -15.901 },
-    { longitude: -48.053, latitude: -15.893 },
-    { longitude: -48.058, latitude: -15.882 },
-    { longitude: -48.058, latitude: -15.878 },
-    { longitude: -48.056, latitude: -15.877 },
-];
-
-export const coordinatesRiachoFundoII = [
-    { longitude: -48.064, latitude: -15.884 },
-    { longitude: -48.063, latitude: -15.879 },
-    { longitude: -48.056, latitude: -15.876 },
-    { longitude: -48.052, latitude: -15.876 },
-    { longitude: -48.046, latitude: -15.889 },
-    { longitude: -48.042, latitude: -15.891 },
-    { longitude: -48.041, latitude: -15.894 },
-    { longitude: -48.034, latitude: -15.895 },
-    { longitude: -48.028, latitude: -15.891 },
-    { longitude: -48.02, latitude: -15.893 },
-    { longitude: -48.011, latitude: -15.902 },
-    { longitude: -48.007, latitude: -15.903 },
-    { longitude: -48.004, latitude: -15.907 },
-    { longitude: -48.002, latitude: -15.907 },
-    { longitude: -47.998, latitude: -15.912 },
-    { longitude: -47.997, latitude: -15.916 },
-    { longitude: -47.994, latitude: -15.918 },
-    { longitude: -47.995, latitude: -15.924 },
-    { longitude: -47.999, latitude: -15.927 },
-    { longitude: -48.001, latitude: -15.932 },
-    { longitude: -48.005, latitude: -15.936 },
-    { longitude: -47.996, latitude: -15.945 },
-    { longitude: -47.996, latitude: -15.949 },
-    { longitude: -48.019, latitude: -15.97 },
-    { longitude: -48.023, latitude: -15.97 },
-    { longitude: -48.044, latitude: -15.945 },
-    { longitude: -48.064, latitude: -15.884 },
-];
-
-export const coordinatesSamambaia = [
-    { longitude: -48.263, latitude: -15.921 },
-    { longitude: -48.262, latitude: -15.916 },
-    { longitude: -48.259, latitude: -15.914 },
-    { longitude: -48.259, latitude: -15.91 },
-    { longitude: -48.255, latitude: -15.906 },
-    { longitude: -48.245, latitude: -15.902 },
-    { longitude: -48.241, latitude: -15.899 },
-    { longitude: -48.24, latitude: -15.896 },
-    { longitude: -48.234, latitude: -15.895 },
-    { longitude: -48.232, latitude: -15.897 },
-    { longitude: -48.224, latitude: -15.897 },
-    { longitude: -48.218, latitude: -15.893 },
-    { longitude: -48.212, latitude: -15.892 },
-    { longitude: -48.208, latitude: -15.888 },
-    { longitude: -48.205, latitude: -15.888 },
-    { longitude: -48.204, latitude: -15.882 },
-    { longitude: -48.202, latitude: -15.881 },
-    { longitude: -48.196, latitude: -15.869 },
-    { longitude: -48.188, latitude: -15.864 },
-    { longitude: -48.185, latitude: -15.857 },
-    { longitude: -48.181, latitude: -15.857 },
-    { longitude: -48.18, latitude: -15.854 },
-    { longitude: -48.166, latitude: -15.848 },
-    { longitude: -48.161, latitude: -15.848 },
-    { longitude: -48.157, latitude: -15.851 },
-    { longitude: -48.149, latitude: -15.85 },
-    { longitude: -48.14, latitude: -15.858 },
-    { longitude: -48.135, latitude: -15.86 },
-    { longitude: -48.131, latitude: -15.859 },
-    { longitude: -48.126, latitude: -15.862 },
-    { longitude: -48.119, latitude: -15.862 },
-    { longitude: -48.114, latitude: -15.864 },
-    { longitude: -48.113, latitude: -15.859 },
-    { longitude: -48.108, latitude: -15.854 },
-    { longitude: -48.102, latitude: -15.853 },
-    { longitude: -48.095, latitude: -15.849 },
-    { longitude: -48.091, latitude: -15.849 },
-    { longitude: -48.086, latitude: -15.844 },
-    { longitude: -48.08, latitude: -15.842 },
-    { longitude: -48.079, latitude: -15.84 },
-    { longitude: -48.073, latitude: -15.84 },
-    { longitude: -48.068, latitude: -15.845 },
-    { longitude: -48.067, latitude: -15.848 },
-    { longitude: -48.065, latitude: -15.848 },
-    { longitude: -48.061, latitude: -15.855 },
-    { longitude: -48.052, latitude: -15.853 },
-    { longitude: -48.05, latitude: -15.856 },
-    { longitude: -48.048, latitude: -15.855 },
-    { longitude: -48.045, latitude: -15.856 },
-    { longitude: -48.043, latitude: -15.859 },
-    { longitude: -48.037, latitude: -15.86 },
-    { longitude: -48.033, latitude: -15.867 },
-    { longitude: -48.033, latitude: -15.87 },
-    { longitude: -48.035, latitude: -15.872 },
-    { longitude: -48.054, latitude: -15.885 },
-    { longitude: -48.16, latitude: -15.917 },
-    { longitude: -48.162, latitude: -15.925 },
-    { longitude: -48.164, latitude: -15.927 },
-    { longitude: -48.17, latitude: -15.927 },
-    { longitude: -48.175, latitude: -15.922 },
-    { longitude: -48.179, latitude: -15.927 },
-    { longitude: -48.182, latitude: -15.928 },
-    { longitude: -48.189, latitude: -15.924 },
-    { longitude: -48.199, latitude: -15.928 },
-    { longitude: -48.211, latitude: -15.936 },
-    { longitude: -48.223, latitude: -15.941 },
-    { longitude: -48.228, latitude: -15.941 },
-    { longitude: -48.237, latitude: -15.946 },
-    { longitude: -48.249, latitude: -15.947 },
-    { longitude: -48.256, latitude: -15.946 },
-    { longitude: -48.259, latitude: -15.936 },
-    { longitude: -48.263, latitude: -15.933 },
-    { longitude: -48.263, latitude: -15.921 },
-];
-
-export const coordinatesSantaMaria = [
-    { longitude: -48.057, latitude: -16.034 },
-    { longitude: -48.053, latitude: -16.028 },
-    { longitude: -48.053, latitude: -16.025 },
-    { longitude: -48.05, latitude: -16.023 },
-    { longitude: -48.048, latitude: -16.017 },
-    { longitude: -48.041, latitude: -16.016 },
-    { longitude: -48.033, latitude: -16.008 },
-    { longitude: -48.016, latitude: -15.996 },
-    { longitude: -48.014, latitude: -15.989 },
-    { longitude: -48.011, latitude: -15.986 },
-    { longitude: -48.008, latitude: -15.975 },
-    { longitude: -48.006, latitude: -15.974 },
-    { longitude: -47.997, latitude: -15.974 },
-    { longitude: -47.973, latitude: -15.97 },
-    { longitude: -47.942, latitude: -15.98 },
-    { longitude: -47.934, latitude: -15.98 },
-    { longitude: -47.927, latitude: -15.978 },
-    { longitude: -47.879, latitude: -15.96 },
-    { longitude: -47.835, latitude: -15.924 },
-    { longitude: -47.831, latitude: -15.924 },
-    { longitude: -47.829, latitude: -15.926 },
-    { longitude: -47.814, latitude: -15.959 },
-    { longitude: -47.816, latitude: -15.982 },
-    { longitude: -47.812, latitude: -15.986 },
-    { longitude: -47.797, latitude: -16.011 },
-    { longitude: -47.798, latitude: -16.025 },
-    { longitude: -47.813, latitude: -16.054 },
-    { longitude: -47.941, latitude: -16.053 },
-    { longitude: -48.04, latitude: -16.054 },
-    { longitude: -48.05, latitude: -16.04 },
-    { longitude: -48.056, latitude: -16.041 },
-    { longitude: -48.058, latitude: -16.039 },
-    { longitude: -48.057, latitude: -16.034 },
-];
-
-export const coordinatesSaoSebastiao = [
-    { longitude: -47.836, latitude: -15.93 },
-    { longitude: -47.837, latitude: -15.927 },
-    { longitude: -47.824, latitude: -15.905 },
-    { longitude: -47.823, latitude: -15.898 },
-    { longitude: -47.826, latitude: -15.884 },
-    { longitude: -47.823, latitude: -15.881 },
-    { longitude: -47.815, latitude: -15.881 },
-    { longitude: -47.8, latitude: -15.886 },
-    { longitude: -47.8, latitude: -15.882 },
-    { longitude: -47.804, latitude: -15.879 },
-    { longitude: -47.805, latitude: -15.876 },
-    { longitude: -47.796, latitude: -15.867 },
-    { longitude: -47.79, latitude: -15.865 },
-    { longitude: -47.783, latitude: -15.869 },
-    { longitude: -47.783, latitude: -15.876 },
-    { longitude: -47.775, latitude: -15.878 },
-    { longitude: -47.772, latitude: -15.881 },
-    { longitude: -47.766, latitude: -15.879 },
-    { longitude: -47.76, latitude: -15.882 },
-    { longitude: -47.757, latitude: -15.886 },
-    { longitude: -47.755, latitude: -15.881 },
-    { longitude: -47.756, latitude: -15.872 },
-    { longitude: -47.754, latitude: -15.868 },
-    { longitude: -47.751, latitude: -15.865 },
-    { longitude: -47.74, latitude: -15.864 },
-    { longitude: -47.738, latitude: -15.867 },
-    { longitude: -47.726, latitude: -15.871 },
-    { longitude: -47.721, latitude: -15.871 },
-    { longitude: -47.718, latitude: -15.873 },
-    { longitude: -47.709, latitude: -15.871 },
-    { longitude: -47.706, latitude: -15.873 },
-    { longitude: -47.704, latitude: -15.877 },
-    { longitude: -47.704, latitude: -15.882 },
-    { longitude: -47.706, latitude: -15.884 },
-    { longitude: -47.705, latitude: -15.888 },
-    { longitude: -47.699, latitude: -15.891 },
-    { longitude: -47.693, latitude: -15.9 },
-    { longitude: -47.69, latitude: -15.9 },
-    { longitude: -47.685, latitude: -15.903 },
-    { longitude: -47.684, latitude: -15.906 },
-    { longitude: -47.681, latitude: -15.906 },
-    { longitude: -47.677, latitude: -15.909 },
-    { longitude: -47.676, latitude: -15.912 },
-    { longitude: -47.678, latitude: -15.915 },
-    { longitude: -47.677, latitude: -15.918 },
-    { longitude: -47.672, latitude: -15.922 },
-    { longitude: -47.672, latitude: -15.924 },
-    { longitude: -47.666, latitude: -15.927 },
-    { longitude: -47.666, latitude: -15.931 },
-    { longitude: -47.664, latitude: -15.933 },
-    { longitude: -47.665, latitude: -15.939 },
-    { longitude: -47.662, latitude: -15.943 },
-    { longitude: -47.664, latitude: -15.947 },
-    { longitude: -47.663, latitude: -15.949 },
-    { longitude: -47.659, latitude: -15.95 },
-    { longitude: -47.656, latitude: -15.953 },
-    { longitude: -47.656, latitude: -15.958 },
-    { longitude: -47.658, latitude: -15.96 },
-    { longitude: -47.652, latitude: -15.965 },
-    { longitude: -47.65, latitude: -15.965 },
-    { longitude: -47.644, latitude: -15.959 },
-    { longitude: -47.637, latitude: -15.955 },
-    { longitude: -47.632, latitude: -15.944 },
-    { longitude: -47.62, latitude: -15.935 },
-    { longitude: -47.607, latitude: -15.933 },
-    { longitude: -47.603, latitude: -15.937 },
-    { longitude: -47.584, latitude: -15.967 },
-    { longitude: -47.584, latitude: -15.973 },
-    { longitude: -47.611, latitude: -16.016 },
-    { longitude: -47.614, latitude: -16.049 },
-    { longitude: -47.613, latitude: -16.051 },
-    { longitude: -47.615, latitude: -16.054 },
-    { longitude: -47.816, latitude: -16.054 },
-    { longitude: -47.818, latitude: -16.053 },
-    { longitude: -47.819, latitude: -16.049 },
-    { longitude: -47.806, latitude: -16.023 },
-    { longitude: -47.805, latitude: -16.012 },
-    { longitude: -47.818, latitude: -15.991 },
-    { longitude: -47.824, latitude: -15.984 },
-    { longitude: -47.822, latitude: -15.96 },
-    { longitude: -47.836, latitude: -15.93 },
-];
-
-export const coordinatesSobradinhoI = [
-    { longitude: -48.036, latitude: -15.559 },
-    { longitude: -48.027, latitude: -15.553 },
-    { longitude: -48.018, latitude: -15.55 },
-    { longitude: -48.017, latitude: -15.548 },
-    { longitude: -48.013, latitude: -15.547 },
-    { longitude: -48.003, latitude: -15.54 },
-    { longitude: -47.997, latitude: -15.538 },
-    { longitude: -47.99, latitude: -15.528 },
-    { longitude: -47.981, latitude: -15.525 },
-    { longitude: -47.976, latitude: -15.525 },
-    { longitude: -47.972, latitude: -15.529 },
-    { longitude: -47.963, latitude: -15.533 },
-    { longitude: -47.947, latitude: -15.548 },
-    { longitude: -47.947, latitude: -15.56 },
-    { longitude: -47.952, latitude: -15.566 },
-    { longitude: -47.95, latitude: -15.57 },
-    { longitude: -47.946, latitude: -15.569 },
-    { longitude: -47.943, latitude: -15.571 },
-    { longitude: -47.937, latitude: -15.569 },
-    { longitude: -47.933, latitude: -15.576 },
-    { longitude: -47.928, latitude: -15.579 },
-    { longitude: -47.924, latitude: -15.584 },
-    { longitude: -47.902, latitude: -15.59 },
-    { longitude: -47.897, latitude: -15.586 },
-    { longitude: -47.887, latitude: -15.589 },
-    { longitude: -47.884, latitude: -15.592 },
-    { longitude: -47.884, latitude: -15.595 },
-    { longitude: -47.881, latitude: -15.6 },
-    { longitude: -47.877, latitude: -15.601 },
-    { longitude: -47.875, latitude: -15.603 },
-    { longitude: -47.875, latitude: -15.606 },
-    { longitude: -47.872, latitude: -15.608 },
-    { longitude: -47.872, latitude: -15.619 },
-    { longitude: -47.865, latitude: -15.623 },
-    { longitude: -47.863, latitude: -15.626 },
-    { longitude: -47.859, latitude: -15.625 },
-    { longitude: -47.855, latitude: -15.627 },
-    { longitude: -47.852, latitude: -15.633 },
-    { longitude: -47.85, latitude: -15.633 },
-    { longitude: -47.847, latitude: -15.636 },
-    { longitude: -47.843, latitude: -15.646 },
-    { longitude: -47.838, latitude: -15.639 },
-    { longitude: -47.835, latitude: -15.638 },
-    { longitude: -47.831, latitude: -15.639 },
-    { longitude: -47.829, latitude: -15.641 },
-    { longitude: -47.827, latitude: -15.65 },
-    { longitude: -47.821, latitude: -15.65 },
-    { longitude: -47.818, latitude: -15.646 },
-    { longitude: -47.824, latitude: -15.643 },
-    { longitude: -47.829, latitude: -15.631 },
-    { longitude: -47.829, latitude: -15.625 },
-    { longitude: -47.827, latitude: -15.621 },
-    { longitude: -47.822, latitude: -15.618 },
-    { longitude: -47.823, latitude: -15.614 },
-    { longitude: -47.829, latitude: -15.609 },
-    { longitude: -47.829, latitude: -15.605 },
-    { longitude: -47.826, latitude: -15.602 },
-    { longitude: -47.822, latitude: -15.602 },
-    { longitude: -47.815, latitude: -15.606 },
-    { longitude: -47.812, latitude: -15.598 },
-    { longitude: -47.809, latitude: -15.596 },
-    { longitude: -47.8, latitude: -15.594 },
-    { longitude: -47.78, latitude: -15.593 },
-    { longitude: -47.772, latitude: -15.597 },
-    { longitude: -47.748, latitude: -15.629 },
-    { longitude: -47.738, latitude: -15.674 },
-    { longitude: -47.738, latitude: -15.677 },
-    { longitude: -47.743, latitude: -15.681 },
-    { longitude: -47.743, latitude: -15.684 },
-    { longitude: -47.738, latitude: -15.686 },
-    { longitude: -47.736, latitude: -15.69 },
-    { longitude: -47.734, latitude: -15.69 },
-    { longitude: -47.731, latitude: -15.694 },
-    { longitude: -47.715, latitude: -15.693 },
-    { longitude: -47.714, latitude: -15.695 },
-    { longitude: -47.71, latitude: -15.696 },
-    { longitude: -47.697, latitude: -15.695 },
-    { longitude: -47.689, latitude: -15.699 },
-    { longitude: -47.685, latitude: -15.699 },
-    { longitude: -47.677, latitude: -15.703 },
-    { longitude: -47.674, latitude: -15.703 },
-    { longitude: -47.669, latitude: -15.708 },
-    { longitude: -47.671, latitude: -15.719 },
-    { longitude: -47.67, latitude: -15.731 },
-    { longitude: -47.672, latitude: -15.735 },
-    { longitude: -47.677, latitude: -15.738 },
-    { longitude: -47.682, latitude: -15.738 },
-    { longitude: -47.688, latitude: -15.733 },
-    { longitude: -47.69, latitude: -15.733 },
-    { longitude: -47.692, latitude: -15.73 },
-    { longitude: -47.698, latitude: -15.732 },
-    { longitude: -47.72, latitude: -15.73 },
-    { longitude: -47.729, latitude: -15.732 },
-    { longitude: -47.732, latitude: -15.731 },
-    { longitude: -47.743, latitude: -15.732 },
-    { longitude: -47.747, latitude: -15.728 },
-    { longitude: -47.751, latitude: -15.727 },
-    { longitude: -47.755, latitude: -15.722 },
-    { longitude: -47.758, latitude: -15.722 },
-    { longitude: -47.762, latitude: -15.719 },
-    { longitude: -47.767, latitude: -15.725 },
-    { longitude: -47.772, latitude: -15.728 },
-    { longitude: -47.778, latitude: -15.729 },
-    { longitude: -47.78, latitude: -15.728 },
-    { longitude: -47.791, latitude: -15.732 },
-    { longitude: -47.796, latitude: -15.725 },
-    { longitude: -47.836, latitude: -15.698 },
-    { longitude: -47.851, latitude: -15.697 },
-    { longitude: -47.862, latitude: -15.692 },
-    { longitude: -47.886, latitude: -15.671 },
-    { longitude: -48.006, latitude: -15.582 },
-    { longitude: -48.009, latitude: -15.582 },
-    { longitude: -48.015, latitude: -15.586 },
-    { longitude: -48.019, latitude: -15.586 },
-    { longitude: -48.027, latitude: -15.576 },
-    { longitude: -48.039, latitude: -15.567 },
-    { longitude: -48.039, latitude: -15.562 },
-    { longitude: -48.036, latitude: -15.559 },
-];
-
-export const coordinatesSobradinhoII = [
-    { longitude: -48.039, latitude: -15.533 },
-    { longitude: -48.043, latitude: -15.527 },
-    { longitude: -48.044, latitude: -15.517 },
-    { longitude: -48.038, latitude: -15.504 },
-    { longitude: -48.04, latitude: -15.502 },
-    { longitude: -48.039, latitude: -15.497 },
-    { longitude: -48.037, latitude: -15.496 },
-    { longitude: -47.77, latitude: -15.496 },
-    { longitude: -47.767, latitude: -15.498 },
-    { longitude: -47.766, latitude: -15.507 },
-    { longitude: -47.767, latitude: -15.514 },
-    { longitude: -47.765, latitude: -15.516 },
-    { longitude: -47.76, latitude: -15.516 },
-    { longitude: -47.757, latitude: -15.521 },
-    { longitude: -47.758, latitude: -15.524 },
-    { longitude: -47.753, latitude: -15.525 },
-    { longitude: -47.75, latitude: -15.528 },
-    { longitude: -47.746, latitude: -15.537 },
-    { longitude: -47.746, latitude: -15.541 },
-    { longitude: -47.749, latitude: -15.546 },
-    { longitude: -47.752, latitude: -15.547 },
-    { longitude: -47.751, latitude: -15.548 },
-    { longitude: -47.752, latitude: -15.556 },
-    { longitude: -47.756, latitude: -15.558 },
-    { longitude: -47.759, latitude: -15.562 },
-    { longitude: -47.761, latitude: -15.562 },
-    { longitude: -47.756, latitude: -15.572 },
-    { longitude: -47.756, latitude: -15.579 },
-    { longitude: -47.759, latitude: -15.583 },
-    { longitude: -47.762, latitude: -15.584 },
-    { longitude: -47.763, latitude: -15.589 },
-    { longitude: -47.765, latitude: -15.59 },
-    { longitude: -47.771, latitude: -15.602 },
-    { longitude: -47.773, latitude: -15.604 },
-    { longitude: -47.78, latitude: -15.604 },
-    { longitude: -47.782, latitude: -15.601 },
-    { longitude: -47.805, latitude: -15.603 },
-    { longitude: -47.811, latitude: -15.614 },
-    { longitude: -47.814, latitude: -15.615 },
-    { longitude: -47.812, latitude: -15.619 },
-    { longitude: -47.812, latitude: -15.624 },
-    { longitude: -47.815, latitude: -15.626 },
-    { longitude: -47.821, latitude: -15.626 },
-    { longitude: -47.821, latitude: -15.631 },
-    { longitude: -47.819, latitude: -15.632 },
-    { longitude: -47.817, latitude: -15.638 },
-    { longitude: -47.81, latitude: -15.64 },
-    { longitude: -47.808, latitude: -15.643 },
-    { longitude: -47.813, latitude: -15.654 },
-    { longitude: -47.821, latitude: -15.658 },
-    { longitude: -47.827, latitude: -15.658 },
-    { longitude: -47.831, latitude: -15.657 },
-    { longitude: -47.836, latitude: -15.649 },
-    { longitude: -47.844, latitude: -15.66 },
-    { longitude: -47.848, latitude: -15.66 },
-    { longitude: -47.85, latitude: -15.659 },
-    { longitude: -47.851, latitude: -15.656 },
-    { longitude: -47.85, latitude: -15.649 },
-    { longitude: -47.854, latitude: -15.641 },
-    { longitude: -47.859, latitude: -15.638 },
-    { longitude: -47.86, latitude: -15.634 },
-    { longitude: -47.866, latitude: -15.634 },
-    { longitude: -47.869, latitude: -15.632 },
-    { longitude: -47.871, latitude: -15.625 },
-    { longitude: -47.867, latitude: -15.62 },
-    { longitude: -47.863, latitude: -15.618 },
-    { longitude: -47.864, latitude: -15.613 },
-    { longitude: -47.859, latitude: -15.599 },
-    { longitude: -47.859, latitude: -15.592 },
-    { longitude: -47.854, latitude: -15.584 },
-    { longitude: -47.85, latitude: -15.573 },
-    { longitude: -47.816, latitude: -15.556 },
-    { longitude: -47.817, latitude: -15.55 },
-    { longitude: -47.834, latitude: -15.549 },
-    { longitude: -47.841, latitude: -15.555 },
-    { longitude: -47.865, latitude: -15.561 },
-    { longitude: -47.895, latitude: -15.586 },
-    { longitude: -47.918, latitude: -15.589 },
-    { longitude: -47.932, latitude: -15.587 },
-    { longitude: -47.939, latitude: -15.582 },
-    { longitude: -47.941, latitude: -15.578 },
-    { longitude: -47.952, latitude: -15.58 },
-    { longitude: -47.958, latitude: -15.573 },
-    { longitude: -47.96, latitude: -15.567 },
-    { longitude: -47.959, latitude: -15.562 },
-    { longitude: -47.955, latitude: -15.558 },
-    { longitude: -47.956, latitude: -15.551 },
-    { longitude: -47.961, latitude: -15.545 },
-    { longitude: -47.968, latitude: -15.541 },
-    { longitude: -47.969, latitude: -15.538 },
-    { longitude: -47.978, latitude: -15.536 },
-    { longitude: -47.98, latitude: -15.534 },
-    { longitude: -47.986, latitude: -15.536 },
-    { longitude: -47.99, latitude: -15.543 },
-    { longitude: -48.005, latitude: -15.55 },
-    { longitude: -48.007, latitude: -15.553 },
-    { longitude: -48.013, latitude: -15.555 },
-    { longitude: -48.017, latitude: -15.559 },
-    { longitude: -48.033, latitude: -15.566 },
-    { longitude: -48.037, latitude: -15.566 },
-    { longitude: -48.039, latitude: -15.563 },
-    { longitude: -48.038, latitude: -15.557 },
-    { longitude: -48.044, latitude: -15.554 },
-    { longitude: -48.048, latitude: -15.549 },
-    { longitude: -48.046, latitude: -15.54 },
-    { longitude: -48.039, latitude: -15.533 },
-];
-
-export const coordinatesSIA = [
-    { longitude: -47.995, latitude: -15.797 },
-    { longitude: -47.993, latitude: -15.785 },
-    { longitude: -47.99, latitude: -15.783 },
-    { longitude: -47.982, latitude: -15.785 },
-    { longitude: -47.969, latitude: -15.792 },
-    { longitude: -47.958, latitude: -15.79 },
-    { longitude: -47.947, latitude: -15.783 },
-    { longitude: -47.939, latitude: -15.782 },
-    { longitude: -47.937, latitude: -15.785 },
-    { longitude: -47.944, latitude: -15.805 },
-    { longitude: -47.946, latitude: -15.822 },
-    { longitude: -47.949, latitude: -15.824 },
-    { longitude: -47.954, latitude: -15.824 },
-    { longitude: -47.96, latitude: -15.821 },
-    { longitude: -47.968, latitude: -15.825 },
-    { longitude: -47.972, latitude: -15.825 },
-    { longitude: -47.975, latitude: -15.82 },
-    { longitude: -47.974, latitude: -15.811 },
-    { longitude: -47.979, latitude: -15.809 },
-    { longitude: -47.983, latitude: -15.811 },
-    { longitude: -47.994, latitude: -15.81 },
-    { longitude: -47.996, latitude: -15.808 },
-    { longitude: -47.995, latitude: -15.797 },
-];
-
-export const coordinatesVarjao = [
-    { longitude: -47.894, latitude: -15.704 },
-    { longitude: -47.886, latitude: -15.697 },
-    { longitude: -47.882, latitude: -15.698 },
-    { longitude: -47.879, latitude: -15.702 },
-    { longitude: -47.871, latitude: -15.704 },
-    { longitude: -47.869, latitude: -15.706 },
-    { longitude: -47.866, latitude: -15.712 },
-    { longitude: -47.867, latitude: -15.718 },
-    { longitude: -47.869, latitude: -15.721 },
-    { longitude: -47.874, latitude: -15.722 },
-    { longitude: -47.877, latitude: -15.719 },
-    { longitude: -47.879, latitude: -15.719 },
-    { longitude: -47.882, latitude: -15.715 },
-    { longitude: -47.894, latitude: -15.711 },
-    { longitude: -47.896, latitude: -15.708 },
-    { longitude: -47.894, latitude: -15.704 },
-];
-
-export const coordinatesVicentePires = [
-    { longitude: -48.06, latitude: -15.807 },
-    { longitude: -48.06, latitude: -15.802 },
-    { longitude: -48.056, latitude: -15.79 },
-    { longitude: -48.053, latitude: -15.788 },
-    { longitude: -48.027, latitude: -15.787 },
-    { longitude: -48.005, latitude: -15.782 },
-    { longitude: -47.999, latitude: -15.784 },
-    { longitude: -47.999, latitude: -15.789 },
-    { longitude: -48.002, latitude: -15.792 },
-    { longitude: -48.003, latitude: -15.802 },
-    { longitude: -48.003, latitude: -15.808 },
-    { longitude: -48.001, latitude: -15.812 },
-    { longitude: -48.002, latitude: -15.818 },
-    { longitude: -48.015, latitude: -15.822 },
-    { longitude: -48.032, latitude: -15.833 },
-    { longitude: -48.047, latitude: -15.835 },
-    { longitude: -48.053, latitude: -15.834 },
-    { longitude: -48.059, latitude: -15.816 },
-    { longitude: -48.06, latitude: -15.807 },
-];
-
-export const coordinates = [
-    {
-        name: "Gama",
-        color: "rgba(255, 255, 255, 0)",
-        coordinates: coordinatesGama,
-    },
-    {
-        name: "Guará",
-        color: "rgba(0, 255, 0, 0)",
-        coordinates: coordinatesGuara,
-    },
-    {
-        name: "Taguatinga",
-        color: "rgba(0, 0, 255, 0)",
-        coordinates: coordinatesTaguatinga,
-    },
-    {
-        name: "Águas Claras",
-        color: "rgba(255, 255, 0, 0)",
-        coordinates: coordinatesAguasClaras,
-    },
-    {
-        name: "Ceilândia",
-        color: "rgba(0, 255, 255, 0)",
-        coordinates: coordinatesCeilandia,
-    },
-    {
-        name: "Brasília",
-        color: "rgba(255, 0, 0, 0)",
-        coordinates: coordinatesBrasilia,
-    },
-    {
-        name: "Sudoeste Octogonal",
-        color: "rgba(0, 0, 255, 0)",
-        coordinates: coordinatesSudoesteOctogonal,
-    },
-    {
-        name: "Brazlândia",
-        color: "rgba(0, 255, 0, 0)",
-        coordinates: coordinatesBrazlandia,
-    },
-    {
-        name: "Candangolândia",
-        color: "rgba(184, 63, 191, 0)",
-        coordinates: coordinatesCandangolandia,
-    },
-    {
-        name: "Cruzeiro",
-        color: "rgba(184, 63, 191, 0)",
-        coordinates: coordinatesCruzeiro,
-    },
-    {
-        name: "Estrutural SCIA",
-        color: "rgba(0, 0, 255, 0)",
-        coordinates: coordinatesEstruturalSCIA,
-    },
-    {
-        name: "Fercal",
-        color: "rgba(191, 131, 50, 0)",
-        coordinates: coordinatesFercal,
-    },
-    {
-        name: "Itapoã",
-        color: "rgba(0, 0, 255, 0)",
-        coordinates: coordinatesItapoa,
-    },
-    {
-        name: "Jardim Botânico",
-        color: "rgba(0, 0, 255, 0)",
-        coordinates: coordinatesJardimBotanico,
-    },
-    {
-        name: "Lago Norte",
-        color: "rgba(0, 255, 255, 0)",
-        coordinates: coordinatesLagoNorte,
-    },
-    {
-        name: "Lago Sul",
-        color: "rgba(0, 255, 0, 0)",
-        coordinates: coordinatesLagoSul,
-    },
-    {
-        name: "Núcleo Bandeirante",
-        color: "rgba(191, 131, 50, 0)",
-        coordinates: coordinatesNucleoBandeirante,
-    },
-    {
-        name: "Paranoá",
-        color: "rgba(184, 63, 191, 0)",
-        coordinates: coordinatesParanoa,
-    },
-    {
-        name: "Park Way",
-        color: "rgba(0, 0, 255, 0)",
-        coordinates: coordinatesParkWay,
-    },
-    {
-        name: "Planaltina",
-        color: "rgba(255, 0, 0, 0)",
-        coordinates: coordinatesPlanaltina,
-    },
-    {
-        name: "Recanto das Emas",
-        color: "rgba(0, 255, 0, 0)",
-        coordinates: coordinatesRecantoDasEmas,
-    },
-    {
-        name: "Riacho Fundo I",
-        color: "rgba(184, 63, 191, 0)",
-        coordinates: coordinatesRiachoFundoI,
-    },
-    {
-        name: "Riacho Fundo II",
-        color: "rgba(0, 255, 255, 0)",
-        coordinates: coordinatesRiachoFundoII,
-    },
-    {
-        name: "Samambaia",
-        color: "rgba(0, 0, 50, 0)",
-        coordinates: coordinatesSamambaia,
-    },
-    {
-        name: "Santa Maria",
-        color: "rgba(0, 0, 50, 0)",
-        coordinates: coordinatesSantaMaria,
-    },
-    {
-        name: "São Sebastião",
-        color: "rgba(255, 0, 0, 0)",
-        coordinates: coordinatesSaoSebastiao,
-    },
-    {
-        name: "Sobradinho I",
-        color: "rgba(255, 255, 0, 0)",
-        coordinates: coordinatesSobradinhoI,
-    },
-    {
-        name: "Sobradinho II",
-        color: "rgba(184, 63, 191, 0)",
-        coordinates: coordinatesSobradinhoII,
-    },
-    {
-        name: "SIA",
-        color: "rgba(0, 255, 255, 0)",
-        coordinates: coordinatesSIA,
-    },
-    {
-        name: "Varjão",
-        color: "rgba(0, 0, 0, 0)",
-        coordinates: coordinatesVarjao,
-    },
-    {
-        name: "Vicente Pires",
-        color: "rgba(0, 0, 0, 0)",
-        coordinates: coordinatesVicentePires,
-    },
-];
